bugid-51482: replacing previous VET with new directory, removing VideoEmbedTool2 dir
