--- extensions/wikia/VideoEmbedTool2/VideoEmbedTool.i18n.php
@@ -1,2649 +1,0 @@
-<?php
-
-$messages = array();
-
-$messages['en'] = array(
-	'vet-upload' => 'Full URL',
-	'vet-notlogged' => 'Log in or register to insert videos',
-	'vet-find' => 'Find a video',
-	'vet-thiswiki' => 'This wiki',
-	'vet-video-wiki' => 'Wikia Video Library',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Add',
-	'vet-find-btn' => 'Search',
-	'vet-flickr-inf' => 'Search properly licensed free photos from Flickr and import them into your wiki.',
-	'vet-conflict-inf' => '<h1>Oops, a video with that name already exists.</h1>What do you want to do about <b>$1</b>?',
-	'vet-rename' => 'Rename video',
-	'vet-existing' => 'Use existing video',
-	'vet-overwrite' => 'Overwrite the existing video with yours',
-	'vet-insert' => 'Insert',
-	'vet-details-inf' => '<h1>File details</h1>Specify a name for this video. Make it something descriptive.',
-	'vet-name' => 'Name',
-	'vet-license' => 'License',
-	'vet-details-inf2' => 'Video display options',
-	'vet-size' => 'Size',
-	'vet-thumbnail' => 'Thumbnail',
-	'vet-fullsize' => 'Full-size ($1x$2px)',
-	'vet-width' => 'Width',
-	'vet-layout' => 'Position',
-	'vet-style' => 'Style',
- 	'vet-caption' => 'Caption',
- 	'vet-no-caption' => 'No caption',
-	'vet-optional' => ' (Optional)',
-	'vet-insert2' => 'Insert video',
-	'vet-update' => 'Update video',
-	'vet-video-replace-link' => 'To replace this video, please go to [[:$1]]',
-	'vet-prev' => 'Previous 8',
-	'vet-next' => 'Next 8',
-	'vet-metacafe' => 'Metacafe videos ($1 {{PLURAL:$1|result|results}})',
-	'vet-insert3' => 'Insert this video',
-	'vet-recent-inf' => 'Recently inserted videos',
-	'vet-thiswiki2' => 'This wiki videos ($1 {{PLURAL:$1|result|results}})', // @todo FIXME: is template 'results_thiswiki' missing? Cannot find this message being used.
-	'vet-success' => '<h1>Success</h1>',
-	'vet-return' => 'Return to editing',
-	'vet-back' => 'Back',
-	'vet-close' => 'Close',
-	'vet-warn1' => 'You need to specify query first!',
-	'vet-warn2' => 'Please supply URL of the video you like to add',
-	'vet-warn3' => 'Please supply name of the video you like to add',
-	'vet-insert-error' => 'Sorry, something went wrong while inserting this video to the article',
-	'vet-imagelink' => '<b>New!</b> Insert videos',
-	'vet-imagebutton' => 'Insert videos' ,
-	'vet-license-cc' => 'Licenced under Creative Commons Attribution-Share Alike 3.0',
-	'vet-name-incorrect' => 'This video name contains invalid characters, like #',
-	'vet-uploadtext' => '\'\'\'This form allows you to quickly upload and display files in articles.\'\'\'
-
-\'\'See also the [[Special:Log/upload|upload log]] and a [[Special:ImageList|list of all uploaded files]].\'\'',
-	'vet-show-message' => 'show message',
-	'vet-hide-message' => 'hide message',
-	'vet-hide-license-msg' => 'hide license text',
-	'vet-show-license-msg' => 'show license text',
-	'vet-max-thumb' => 'maximum thumbnail size exceeded, reverting to original size',
-	'vet-notallowed' => 'You cannot insert videos on this wiki',
-	'vet-title' => 'Add a video to your article',
-	'vet-description' => 'Enter the full URL from any of the supported sites. See',
-	'vet-supported' => 'Supported video sites:',
-	'vet-preview' => '<i>This preview is not to scale</i>',
-	'vet-bad-url' => 'The supplied URL is invalid',
-	'vet-non-existing' => 'The supplied video does not exist',
-	'vet-main-info' => '<i>Enter the full URL from any of the supported sites</i>',
-	'vet-page-success' => 'Video page was successfully updated. You may have to purge the video page to see the change.',
-	'vet-add-button' => 'Add video',
-	'vet-added-from-gallery' => 'video added from video gallery',
-	'vet-see-all' => 'all',
-	'vet-bad-search' => 'Error loading search query.',
-	'vet-no-preview' => 'You cannot add video from article in preview mode.',
-	'vet-single-success' => 'Video was successfully placed in this article.',
-	'vet-gallery-create-success' => 'A video gallery was added to this article.',
-	'vet-gallery-add-success' => 'Video has been added to the video gallery in this article.',
-	'vet-left' => 'Left',
-	'vet-right' => 'Right',
-	'vet-center' => 'Center',
-	'vet-gallery' => 'Gallery',
-	'vet-protected' => 'This video is protected.',
-	'vet-suggestions' => 'Suggested videos for this article',
-	'vet-back-to-suggestions' => 'Back to suggestions',
-	'vet-url-label' => 'URL',
-	'vet-search-label' => 'Find',
-	'vet-search-placeholder' => 'Search',
-	'vet-add-from-preview' => 'Add video',
-	'vet-no-results-found' => 'No results found',
-	'vet-search-results-WVL' => 'We found $1 results for "$2" on Wikia Video Library',
-	'vet-search-results-local' => 'We found $1 results for "$2" on this Wiki',
-	'vet-search-filter-caption' => 'Search filter',
-
-);
-
-/** Message documentation (Message documentation)
- * @author Crt
- * @author EugeneZelenko
- * @author LWChris
- * @author McDutchie
- * @author Siebrand
- * @author mech
- */
-$messages['qqq'] = array(
-	'vet-upload-btn' => '{{Identical|Next}}Upload button text.',
-	'vet-find-btn' => '{{Identical|Search}}',
-	'vet-flickr-inf' => ' Maybe "free videos" is meant here?',
-	'vet-insert' => '{{Identical|Insert}}',
-	'vet-name' => '{{Identical|Name}}',
-	'vet-license' => '{{Identical|License}}',
-	'vet-size' => '{{Identical|Size}}',
-	'vet-thumbnail' => '{{Identical|Thumbnail}}',
-	'vet-width' => '{{Identical|Width}}',
-	'vet-layout' => '{{Identical|Layout}}',
-	'vet-caption' => '{{Identical|Caption}}',
-	'vet-optional' => '{{Identical|Optional}}',
-	'vet-prev' => '{{Identical|Previous}}',
-	'vet-next' => '{{Identical|Next}}',
-	'vet-metacafe' => '',
-	'vet-thiswiki2' => '',
-	'vet-back' => '{{Identical|Back}}',
-	'vet-close' => '{{Identical|Close}}',
-	'vet-left' => '{{Identical|Left}}',
-	'vet-right' => '{{Identical|Right}}',
-	'vet-gallery' => '{{Identical|Gallery}}',
-	'vet-suggestions' => 'New feature that will be released soon. Header for displaying suggestions for the article being edited',
-	'vet-back-to-suggestions' => 'New feature that will be released soon. Label for a link displayed about search results carousel',
-	'vet-url-label' => 'New feature that will be released soon. Label for adding url input',
-	'vet-search-label' => 'New feature that will be released soon. Label for the search input + name of search submit button',
-	'vet-search-placeholder' => 'New feature that will be released soon. Placeholder for search input',
-	'vet-add-from-preview' => 'New feature that will be released soon. Add video from preview window',
-	'vet-no-results-found' => 'New feature that will be released soon. Info that no results were found for search',
-	'vet-search-results-WVL' => 'New feature that will be released soon. Number of results found for keyword in Wikia Video Library. $1 = number, $2 = keyword',
-	'vet-search-results-local' => 'New feature that will be released soon. Number of results found for keyword on local wiki. $1 = number, $2 = keyword',
-	'vet-center' => 'New feature that will be released soon. Label for video center position layout option',
-	'vet-style' => 'New feature that will be released soon. Style label in display options (plain image or with border and caption)',
-	'vet-no-caption' => 'New feature that will be released soon. Info that no caption is available for plain image style option',
-	'vet-video-wiki' => 'dropdown label for search filter - Wikia Video Library'
-);
-
-/** Abkhazian (Аҧсшәа)
- * @author Temuri rajavi
- */
-$messages['ab'] = array(
-	'vet-caption' => 'Анапынҵамҭа',
-);
-
-/** Afrikaans (Afrikaans)
- * @author Naudefj
- */
-$messages['af'] = array(
-	'vet-upload' => 'Volledige URL',
-	'vet-notlogged' => "Teken aan of registreer om video's by te voeg",
-	'vet-find' => "Vind 'n video",
-	'vet-thiswiki' => 'Hierdie wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Volgende',
-	'vet-find-btn' => 'Soek',
-	'vet-flickr-inf' => "Soek korrek gelisensieerde vrye foto's vanaf Flickr en laai hulle in u wiki in.",
-	'vet-conflict-inf' => "<h1>Oeps, 'n video met daardie naam bestaan reeds.</h1>Wat wil u met <b>$1</b> doen?",
-	'vet-rename' => 'Hernoem video',
-	'vet-existing' => 'Gebruik bestaande video',
-	'vet-overwrite' => 'Oorskryf die bestaande video met u video',
-	'vet-insert' => 'Voeg by',
-	'vet-details-inf' => "<h1>Lêerbesonderhede</h1>
-Verskaf 'n beskrywende naam vir die video.",
-	'vet-name' => 'Naam',
-	'vet-license' => 'Lisensie',
-	'vet-details-inf2' => 'Videovertoon-opsies',
-	'vet-size' => 'Grootte',
-	'vet-thumbnail' => 'Duimnael',
-	'vet-fullsize' => 'Volledige grootte ($1x$2px)',
-	'vet-width' => 'Breedte',
-	'vet-layout' => 'Uitleg',
-	'vet-caption' => 'Byskrif',
-	'vet-optional' => ' (Opsioneel)',
-	'vet-insert2' => 'Voeg video in',
-	'vet-update' => 'Opdateer video',
-	'vet-video-replace-link' => 'Gaan asseblief na [[:$1]] om hierdie video te vervang',
-	'vet-prev' => 'Vorige 8',
-	'vet-next' => 'Volgende 8',
-	'vet-metacafe' => "Metcafe-video's ($1 {{PLURAL:$1|resultaat|resultate}})",
-	'vet-insert3' => 'Voeg hierdie video in',
-	'vet-recent-inf' => "Onlangs ingevoegde video's",
-	'vet-thiswiki2' => "Video's van hierdie wiki ($1 {{PLURAL:$1|resultaat|resultate}})",
-	'vet-success' => '<h1>Geslaag</h1>',
-	'vet-return' => 'Terug na wysiging',
-	'vet-back' => 'terug',
-	'vet-close' => 'sluit',
-	'vet-warn1' => "U moet eers 'n soekopdrag spesifiseer!",
-	'vet-warn2' => 'Verskaf asseblief URL van die video wat u wil byvoeg',
-	'vet-warn3' => 'Verstrek asseblief naam van die video wat u wil byvoeg',
-	'vet-imagelink' => "<b>Nuut!</b> Voeg video's by",
-	'vet-imagebutton' => "Voeg video's by",
-	'vet-license-cc' => 'Gelisensieer onder die Creative Commons-lisensie Erkenning-Insgelyks Deel 3.0',
-	'vet-name-incorrect' => 'Die naam van hierdie video bevat ongeldige karakters. soos #.',
-	'vet-description' => "Verskaf 'n volledige URL vanaf enige van die ondersteunde webwerwe. Sien",
-	'vet-uploadtext' => "'''Hierdie vorm laat u toe om vinnig lêers te laai en in artikels te plaas.'''
-
-''Sien ook die [[Special:Log/upload|oplaailogboek]] en 'n [[Special:ImageList|lys met alle toegevoegde lêers]].''",
-	'vet-show-message' => 'wys boodskap',
-	'vet-hide-message' => 'versteek boodskap',
-	'vet-hide-license-msg' => 'versteek lisensie-teks',
-	'vet-show-license-msg' => 'wys lisensie-teks',
-	'vet-max-thumb' => 'die maksimum duimnael-grootte is oorskry; word na die oorspronklike afmetings teruggestel',
-	'vet-notallowed' => "U kan nie video's op hierdie wiki bysit nie",
-	'vet-title' => "Voeg 'n video by u artikel",
-	'vet-supported' => 'Ondersteunde video-werwe:',
-	'vet-preview' => '<i>Hierdie voorskou is nie korrek geskaal nie</i>',
-	'vet-bad-url' => 'Die opgegewe URL is ongeldig',
-	'vet-non-existing' => 'Die opgegewe video bestaan nie',
-	'vet-main-info' => "<i>Verskaf 'n volledige URL vanaf een van die ondersteunde webwerwe.</i>",
-	'vet-page-success' => 'Die videobladsy is suksesvol opgedateer.
-U sal moontlik die videobladsy moet opruim (purge) om die verandering te sien.',
-	'vet-add-button' => 'Voeg Video by',
-	'vet-added-from-gallery' => 'video bygevoeg vanuit videogalery',
-	'vet-see-all' => 'alle',
-	'vet-bad-search' => "'n Fout het voorgekom tydens die laai van die soekopdrag.",
-	'vet-no-preview' => "U kan nie 'n video by die bladsy voeg as u met 'n voorskou besig is nie",
-	'vet-single-success' => 'Die video is suksesvol by die bladsy bygevoeg.',
-	'vet-gallery-create-success' => "'n Video-galery is by hierdie bladsy bygevoeg",
-	'vet-gallery-add-success' => 'Die video is bygevoeg by die galery op hierdie bladsy.',
-	'vet-left' => 'Links',
-	'vet-right' => 'Regs',
-	'vet-gallery' => 'Gallery',
-	'vet-protected' => 'Hierdie video is beveilig',
-);
-
-/** Aragonese (Aragonés)
- * @author Juanpabl
- */
-$messages['an'] = array(
-	'vet-insert' => 'Insertar',
-	'vet-thumbnail' => 'Miniatura',
-);
-
-/** Arabic (العربية)
- * @author OsamaK
- */
-$messages['ar'] = array(
-	'vet-upload' => 'المسار الكامل',
-	'vet-thiswiki' => 'هذه الويكي',
-	'vet-flickr' => 'فلكر',
-	'vet-upload-btn' => 'اذهب',
-	'vet-find-btn' => 'ابحث',
-	'vet-insert' => 'أدرج',
-	'vet-name' => 'الاسم',
-	'vet-license' => 'الرخصة',
-	'vet-size' => 'الحجم',
-	'vet-thumbnail' => 'مصغرة',
-	'vet-fullsize' => 'الحجم الكلي ($1x$2px)',
-	'vet-width' => 'العرض',
-	'vet-layout' => 'التخطيط',
-	'vet-caption' => 'التعليق',
-	'vet-optional' => ' (اختياري)',
-	'vet-insert2' => 'أدرج الفيديو',
-	'vet-update' => 'حدّث الفيديو',
-	'vet-close' => 'أغلق',
-	'vet-show-message' => 'أظهر الرسالة',
-	'vet-hide-message' => 'أخفِ الرسالة',
-	'vet-hide-license-msg' => 'أخفِ نص الرخصة',
-	'vet-show-license-msg' => 'أظهر نص الرخصة',
-	'vet-title' => 'إضافة فيديو لمقالك',
-	'vet-supported' => 'مواقع الفيديو المدعومة:',
-);
-
-/** Azerbaijani (azərbaycanca)
- * @author Cekli829
- */
-$messages['az'] = array(
-	'vet-name' => 'Ad',
-	'vet-license' => 'Lisenziya',
-	'vet-size' => 'Həcm',
-	'vet-close' => 'Bağla',
-	'vet-left' => 'Sol',
-	'vet-right' => 'Sağ',
-	'vet-gallery' => 'Qalereya',
-);
-
-/** Belarusian (беларуская)
- * @author Тест
- */
-$messages['be'] = array(
-	'vet-license' => 'Ліцэнзія',
-	'vet-size' => 'Памер',
-);
-
-/** Belarusian (Taraškievica orthography) (‪Беларуская (тарашкевіца)‬)
- * @author EugeneZelenko
- * @author Jim-by
- */
-$messages['be-tarask'] = array(
-	'vet-upload' => 'Поўны URL-адрас',
-	'vet-notlogged' => 'Увайдзіце ў сыстэму альбо зарэгіструйцеся, каб устаўляць відэа',
-	'vet-find' => 'Знайсьці відэа',
-	'vet-thiswiki' => '{{SITENAME}}',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Перайсьці',
-	'vet-find-btn' => 'Пошук',
-	'vet-flickr-inf' => 'Пошук вольных выяваў з адпаведнай ліцэнзіяй на Flickr і імпарт іх ў {{GRAMMAR:вінавальны|{{SITENAME}}}}.',
-	'vet-conflict-inf' => '<h1>Відэа з такой назвай ужо існуе.</h1>Што Вы жадаеце зрабіць з <b>$1</b>?',
-	'vet-rename' => 'Перайменаваць відэа',
-	'vet-existing' => 'Выкарыстоўваць існуючае відэа',
-	'vet-overwrite' => 'Замяніць існуючае відэа Вашым',
-	'vet-insert' => 'Уставіць',
-	'vet-details-inf' => '<h1>Апісаньне файла</h1>Пазначце назву гэтага відэа. Паспрабуйце апісаць ёй зьмест файла.',
-	'vet-name' => 'Назва',
-	'vet-license' => 'Ліцэнзія',
-	'vet-details-inf2' => 'Устаноўкі паказу відэа',
-	'vet-size' => 'Памер',
-	'vet-thumbnail' => 'Мініятура',
-	'vet-fullsize' => 'Поўны памер ($1×$2пкс)',
-	'vet-width' => 'Шырыня',
-	'vet-layout' => 'Разьмяшчэньне',
-	'vet-caption' => 'Подпіс',
-	'vet-optional' => '(Неабавязковае)',
-	'vet-insert2' => 'Уставіць відэа',
-	'vet-update' => 'Абнавіць відэа',
-	'vet-video-replace-link' => 'Каб замяніць гэтае відэа, калі ласка, перайдзіце на [[:$1]]',
-	'vet-prev' => 'Папярэднія 8',
-	'vet-next' => 'Наступныя 8',
-	'vet-metacafe' => 'Відэа з Metacafe ($1 {{PLURAL:$1|вынік|вынікі|вынікаў}})',
-	'vet-insert3' => 'Уставіць гэтае відэа',
-	'vet-recent-inf' => 'Нядаўна дададзеныя відэа',
-	'vet-thiswiki2' => 'Відэа з {{GRAMMAR:родны|{{SITENAME}}}} ($1 {{PLURAL:$1|вынік|вынікі|вынікаў}})',
-	'vet-success' => '<h1>Посьпех</h1>',
-	'vet-return' => 'Вярнуцца да рэдагаваньня',
-	'vet-back' => 'вярнуцца',
-	'vet-close' => 'закрыць',
-	'vet-warn1' => 'Спачатку Вам неабходна пазначыць запыт!',
-	'vet-warn2' => 'Калі ласка, падайце URL-адрас відэа, якое Вы жадаеце дадаць',
-	'vet-warn3' => 'Калі ласка, падайце назву відэа, якое Вы жадаеце дадаць',
-	'vet-imagelink' => '<b>Новае!</b> Уставіць відэа',
-	'vet-imagebutton' => 'Уставіць відэа',
-	'vet-license-cc' => 'На ўмовах ліцэнзіі Creative Commons Attribution-Share Alike 3.0',
-	'vet-name-incorrect' => 'Гэтая назва відэа ўтрымлівае няслушныя сымбалі',
-	'vet-uploadtext' => "'''Гэтая форма дазваляе Вам хутка загружаць і паказваць файлы ў артыкулах.'''
-
-''Глядзіце таксама [[Special:Log/upload|журнал загрузак]] і [[Special:ImageList|сьпіс усіх загружаных файлаў]].''",
-	'vet-show-message' => 'паказаць паведамленьне',
-	'vet-hide-message' => 'схаваць паведамленьне',
-	'vet-hide-license-msg' => 'схаваць тэкст ліцэнзіі',
-	'vet-show-license-msg' => 'паказаць тэкст ліцэнзіі',
-	'vet-max-thumb' => 'перавышаны максымальны памер мініятуры, вярніцеся да пачатковага памеру',
-	'vet-notallowed' => 'Вы ня можаце устаўляць відэа ў {{GRAMMAR:месны|{{SITENAME}}}}.',
-	'vet-title' => 'Дадаць відэа ў Ваш артыкул',
-	'vet-supported' => 'Падтрымлівае відэа-сайты:',
-	'vet-preview' => '<i>Гэты папярэдні прагляд не маштабаваны</i>',
-	'vet-bad-url' => 'Пададзены URL-адрас — няслушны',
-	'vet-non-existing' => 'Пададзенае відэа не існуе',
-	'vet-main-info' => '<i>Знайдзіце відэа альбо ўвядзіце поўны URL-адрас з аднаго з сайтаў, якія падтрымліваюцца</i>',
-	'vet-page-success' => 'Старонка відэа была пасьпяхова абноўленая. Верагодна Вам неабходна ачысьціць кэш старонкі, каб убачыць зьмены.',
-	'vet-add-button' => 'Дадаць відэа',
-	'vet-added-from-gallery' => 'відэа дададзенае з відэа-галерэі',
-	'vet-see-all' => 'паказаць усё',
-	'vet-bad-search' => 'Памылка загрузкі пошукавага запыту.',
-	'vet-no-preview' => 'Вы ня можаце дадаць відэа з артыкула ў рэжыме папярэдняга прагляду',
-	'vet-single-success' => 'Відэа было пасьпяхова разьмешчанае ў гэтым артыкуле',
-	'vet-gallery-create-success' => 'Відэа-галерэя была дададзеная ў гэты артыкул',
-	'vet-gallery-add-success' => 'Відэа было дададзенае ў відэа-галерэю ў гэтым артыкуле',
-	'vet-left' => 'Зьлева',
-	'vet-right' => 'Справа',
-	'vet-gallery' => 'Галерэя',
-	'vet-protected' => 'Гэтае відэа было абароненае',
-);
-
-/** Bulgarian (български)
- * @author DCLXVI
- */
-$messages['bg'] = array(
-	'vet-upload' => 'Пълен адрес',
-	'vet-thiswiki' => 'Това уики',
-	'vet-flickr' => 'Flickr',
-	'vet-find-btn' => 'Търсене',
-	'vet-name' => 'Име',
-	'vet-license' => 'Лиценз',
-	'vet-size' => 'Размер',
-	'vet-fullsize' => 'Пълен размер ($1x$2px)',
-	'vet-width' => 'Ширина',
-	'vet-prev' => 'Предишни 8',
-	'vet-next' => 'Следващи 8',
-	'vet-back' => 'Връщане',
-	'vet-close' => 'Затваряне',
-	'vet-show-message' => 'показване на съобщението',
-	'vet-hide-message' => 'скриване на съобщението',
-	'vet-hide-license-msg' => 'скриване на лицензния текст',
-	'vet-show-license-msg' => 'показване на лицензния текст',
-	'vet-add-button' => 'Добавяне на видео',
-	'vet-see-all' => 'всички',
-	'vet-left' => 'Ляво',
-	'vet-right' => 'Дясно',
-	'vet-gallery' => 'Галерия',
-);
-
-/** Breton (brezhoneg)
- * @author Fulup
- * @author Gwenn-Ael
- * @author Y-M D
- */
-$messages['br'] = array(
-	'vet-upload' => 'URL klok',
-	'vet-notlogged' => "Kevreit pe krouit ur gont implijer evit ensoc'hañ videoioù",
-	'vet-find' => 'Kavout ur video',
-	'vet-thiswiki' => 'Ar Wiki-se',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Mont',
-	'vet-find-btn' => 'Klask',
-	'vet-flickr-inf' => "Klask luc'hskeudennoù gant aotreoù-implijout dieub adal Flcikr hag ensoc'hañ anezho en ho wiki.",
-	'vet-conflict-inf' => "<h1>Ur video gant an anv-se zo dija ! </h1> Petra ho peus c'hoant d'ober gant <b>$1</b> ?",
-	'vet-rename' => 'Adeñvel ar video',
-	'vet-existing' => 'Implijout ur video a zo diouti dija',
-	'vet-overwrite' => "Erlerc'hiañ ar video zo anezhañ gant ho video.",
-	'vet-insert' => "Ensoc'hañ",
-	'vet-details-inf' => "<h1>Anv ar restr</h1>Roit un anv heverk d'ar video, mar plij",
-	'vet-name' => 'Anv',
-	'vet-license' => 'Aotre implijout',
-	'vet-details-inf2' => 'An dibaboù diskwel video',
-	'vet-size' => 'Ment',
-	'vet-thumbnail' => 'Munud',
-	'vet-fullsize' => 'Ment klok ($1x$2px)',
-	'vet-width' => 'Ledander',
-	'vet-layout' => 'Fichañ',
-	'vet-caption' => 'Deskrivadur',
-	'vet-optional' => ' (diret)',
-	'vet-insert2' => "Ensoc'hañ ur Video",
-	'vet-update' => 'Hizivaat ar video',
-	'vet-video-replace-link' => "Evit erlec'hiañ ar video-mañ, kit war-du [[:$1]], mar plij",
-	'vet-prev' => 'an 8 a-raok',
-	'vet-next' => 'an 8 goude',
-	'vet-metacafe' => "Videoioù Metacafe ($1 {{PLURAL:$1|disoc'h|disoc'h}})",
-	'vet-insert3' => "Ensoc'hañ ar video-mañ",
-	'vet-recent-inf' => 'Videoioù ouzhpennet nevez zo',
-	'vet-thiswiki2' => "Videoioù war ar wiki-mañ ($1 {{PLURAL:$1|disoc'h|disoc'h}})",
-	'vet-success' => "<h1>Video ensoc'het</h1>ensoc'het eo ar c'hod da-heul er pennad gant berzh :",
-	'vet-return' => "Distreiñ d'ar c'hemm",
-	'vet-back' => 'distreiñ',
-	'vet-close' => 'serriñ',
-	'vet-warn1' => "Ret eo deoc'h spisaat ho reked da gentañ !",
-	'vet-warn2' => "Ouzhpennit, mar plij, URL ar video hoc'h eus c'hoant da enporzhiañ war {{SITENAME}}",
-	'vet-warn3' => "Roit un anv d'ar video emaoc'h oc'h enporzhiañ, mar plij.",
-	'vet-imagelink' => '<b>Nevez!</b> Ouzhpennit videoioù',
-	'vet-imagebutton' => "Ensoc'hañ videoioù",
-	'vet-license-cc' => "A c'haller kaout dindan aotre-implijout Creative Commons Attribution-Share Alike 3.0",
-	'vet-name-incorrect' => 'Arouezennoù direizh zo en anv ar video-mañ',
-	'vet-uploadtext' => "'''Gant ar furmskrid-mañ e c'hallit enporzhiañ buan restroù ha diskwel anezho er pennadoù.'''
-
-'''Kit da welet ivez '[[Special:Log/upload|istor enporzhiañ ar restroù liesvedia]] ha [[Special:ImageList|roll an holl restroù enporzhiet]].''",
-	'vet-show-message' => 'diskouez ar gemenadenn',
-	'vet-hide-message' => 'kuzhat ar gemenadenn',
-	'vet-hide-license-msg' => 'Kuzhat skrid an aotre-implijout',
-	'vet-show-license-msg' => 'Diskouez skrid an aotre-implijout',
-	'vet-max-thumb' => "Tizhet eo bet ment vrasañ ar skeudennig, distreiñ d'ar vent orin",
-	'vet-notallowed' => "Ne c'hallit ket ensoc'hañ videoioù war ar wiki-mañ",
-	'vet-title' => 'Ouzhpennañ ur video er pennad',
-	'vet-supported' => "Lec'hiennoù videoioù kemeret e karg",
-	'vet-preview' => "<i>n'eo ket ment wirion an alberz</i>",
-	'vet-bad-url' => "Direizh eo ar chomlec'h URL. Gwiriit ar chomlec'h, mar plij",
-	'vet-non-existing' => "N'eus ket eus ar video ouzhpennet",
-	'vet-main-info' => "<i>Klaskit chomlec'h URL a-bezh ur video kemeret e karg war unan eus al lec'hiennoù embreget",
-	'vet-page-success' => "Hizivaet eo bet pajenn ar video gant berzh. Ret eo deoc'h spurjañ pajenn ar video evit gwelet ar c'hemmoù",
-	'vet-add-button' => 'Ouzhpennañ ur Video',
-	'vet-added-from-gallery' => 'Video ouzhpennet adal ar videaoueg',
-	'vet-see-all' => 'gwelet pep tra',
-	'vet-bad-search' => "Ur fazi zo bet pa'z eo bet pellgarget ar reked",
-	'vet-no-preview' => "Ne c'hallit ket ouzhpennañ videoioù er mod rakwelet",
-	'vet-single-success' => 'Lakaet eo bet ar video er pennad ervat.',
-	'vet-gallery-create-success' => 'Ur videaoueg a oa bet ouzhpennet er pennad-mañ',
-	'vet-gallery-add-success' => 'Ur video zo bet ouzhpennet e videaoueg ar pennad-mañ',
-	'vet-left' => 'Kleiz',
-	'vet-right' => 'Dehou',
-	'vet-gallery' => 'Videaoueg',
-	'vet-protected' => 'Gwarezet eo ar video-mañ',
-);
-
-/** Catalan (català)
- * @author Paucabot
- */
-$messages['ca'] = array(
-	'vet-upload' => 'URL completa',
-	'vet-thiswiki' => 'Aquesta Wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-insert' => 'Insereix',
-	'vet-name' => 'Nom',
-	'vet-license' => 'Llicència',
-	'vet-size' => 'Mida',
-	'vet-thumbnail' => 'Miniatura',
-	'vet-width' => 'Amplada',
-	'vet-close' => 'Tanca',
-	'vet-imagebutton' => 'Insereix vídeos',
-	'vet-left' => 'Esquerra',
-	'vet-right' => 'Dreta',
-	'vet-gallery' => 'Galeria',
-	'vet-protected' => 'Aquest vídeo està protegit',
-);
-
-/** Sorani Kurdish (کوردی) */
-$messages['ckb'] = array(
-	'vet-upload-btn' => 'بڕۆ',
-	'vet-find-btn' => 'گەڕان',
-);
-
-/** German (Deutsch)
- * @author Claudia Hattitten
- * @author DaSch
- * @author Geitost
- * @author LWChris
- * @author The Evil IP address
- */
-$messages['de'] = array(
-	'vet-upload' => 'Komplette URL',
-	'vet-notlogged' => 'Melde dich an oder registriere dich, um Videos einzubinden',
-	'vet-find' => 'Video suchen',
-	'vet-thiswiki' => 'Dieses Wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Nächste',
-	'vet-find-btn' => 'Suchen',
-	'vet-flickr-inf' => 'Durchsuche passend lizenzierte Fotos von Flickr und importiere sie in dein Wiki.',
-	'vet-conflict-inf' => '<h1>Es existiert bereits ein Video unter diesem Namen.</h1>Was willst du mit <b>$1</b> machen?',
-	'vet-rename' => 'Video umbenennen',
-	'vet-existing' => 'Existierendes Video verwenden',
-	'vet-overwrite' => 'Das bestehende Video mit deinem überschreiben',
-	'vet-insert' => 'Einfügen',
-	'vet-details-inf' => '<h1>Dateiname</h1>Bitte gib einen möglichst aussagekräftigen Namen für dieses Video an.',
-	'vet-name' => 'Name',
-	'vet-license' => 'Lizenz',
-	'vet-details-inf2' => 'Video Anzeige-Optionen',
-	'vet-size' => 'Größe',
-	'vet-thumbnail' => 'Miniatur',
-	'vet-fullsize' => 'Original ($1x$2px)',
-	'vet-width' => 'Breite',
-	'vet-layout' => 'Ausrichtung',
-	'vet-caption' => 'Beschreibung',
-	'vet-optional' => '(Optional)',
-	'vet-insert2' => 'Video einfügen',
-	'vet-update' => 'Video aktualisieren',
-	'vet-video-replace-link' => 'Um dieses Video zu ersetzten, gehe bitte zu [[:$1]].',
-	'vet-prev' => 'Vorherige 8',
-	'vet-next' => 'Nächste 8',
-	'vet-metacafe' => 'Videos von Metacafe ($1 {{PLURAL:$1|Ergebnis|Ergebnisse}})',
-	'vet-insert3' => 'Dieses Video einfügen',
-	'vet-recent-inf' => 'Zuletzt eingebundene Videos',
-	'vet-thiswiki2' => 'Videos in diesem Wiki ($1 {{PLURAL:$1|Ergebnis|Ergebnisse}})',
-	'vet-success' => '<h1>Erfolg</h1>',
-	'vet-return' => 'Zurück zur Bearbeitung',
-	'vet-back' => 'zurück',
-	'vet-close' => 'schließen',
-	'vet-warn1' => 'Du musst zuerst einen Suchbegriff eingeben!',
-	'vet-warn2' => 'Bitte gib die URL des Videos an, das du einfügen möchtest',
-	'vet-warn3' => 'Bitte gib dem hinzuzufügenden Video einen Namen',
-	'vet-imagelink' => '<b>Neu!</b> Videos hinzufügen',
-	'vet-imagebutton' => 'Videos einfügen',
-	'vet-license-cc' => 'Lizenziert unter der Creative Commons Attribution-Share Alike 3.0 Lizenz',
-	'vet-name-incorrect' => 'Dieser Video-Name enthält ungültige Zeichen, etwa #.',
-	'vet-description' => 'Gib die vollständige URL von einer der unterstützten Websites ein. Siehe',
-	'vet-uploadtext' => "'''Dieses Formular erlaubt dir das schnelle Hochladen und Einbinden von Dateien in Artikel.'''
-
-''Siehe auch das [[Special:Log/upload|Datei-Logbuch]] und die [[Special:ImageList|Liste aller hochgeladenen Dateien]].''",
-	'vet-show-message' => 'Hinweis anzeigen',
-	'vet-hide-message' => 'Hinweis ausblenden',
-	'vet-hide-license-msg' => 'Lizenz-Text ausblenden',
-	'vet-show-license-msg' => 'Lizenz-Text einblenden',
-	'vet-max-thumb' => 'Maximale Vorschau-Größe überschritten, zurückgesetzt auf Original-Größe',
-	'vet-notallowed' => 'Du kannst in diesem Wiki keine Videos einbinden',
-	'vet-title' => 'Füge ein Video in den Artikel ein',
-	'vet-supported' => 'Unterstützte Video-Anbieter:',
-	'vet-preview' => '<i>Dies ist nur eine verkleinerte Vorschau</i>',
-	'vet-bad-url' => 'Die angegebene URL ist ungültig',
-	'vet-non-existing' => 'Das angegebene Video existiert nicht',
-	'vet-main-info' => '<i>Gib die vollständige URL von einer der unterstützten Websites ein</i>',
-	'vet-page-success' => 'Die Video-Seite wurde aktualisiert. Du musst ggf. den Server-Cache per purge leeren um die Änderung zu sehen.',
-	'vet-add-button' => 'Video hinzufügen',
-	'vet-added-from-gallery' => 'Video aus Video-Galerie hinzugefügt',
-	'vet-see-all' => 'alle',
-	'vet-bad-search' => 'Fehler bei der Suchabfrage.',
-	'vet-no-preview' => 'Du kannst aus dem Vorschaumodus kein Video hinzufügen',
-	'vet-single-success' => 'Das Video wurde erfolgreich hinzugefügt',
-	'vet-gallery-create-success' => 'Eine Video-Galerie wurde zum Artikel hinzugefügt',
-	'vet-gallery-add-success' => 'Video wurde zur Video-Galerie im Artikel hinzugefügt',
-	'vet-left' => 'Links',
-	'vet-right' => 'Rechts',
-	'vet-gallery' => 'Galerie',
-	'vet-protected' => 'Dieses Video ist geschützt',
-);
-
-/** German (formal address) (‪Deutsch (Sie-Form)‬)
- * @author LWChris
- * @author Umherirrender
- */
-$messages['de-formal'] = array(
-	'vet-notlogged' => 'Melden Sie sich an oder registrieren Sie sich, um Videos einzubinden',
-	'vet-flickr-inf' => 'Durchsuchen Sie passend lizenzierte Fotos von Flickr und importieren Sie sie in Ihr Wiki.',
-	'vet-conflict-inf' => '<h1>Es existiert bereits ein Video unter diesem Namen.</h1>Was wollen Sie mit <b>$1</b> machen?',
-	'vet-overwrite' => 'Das bestehende Video mit Ihrem überschreiben',
-	'vet-details-inf' => '<h1>Dateiname</h1>Bitte geben Sie einen möglichst aussagekräftigen Namen für dieses Video an.',
-	'vet-fullsize' => 'Original ($1x$2)',
-	'vet-video-replace-link' => 'Um dieses Video zu ersetzten, gehen Sie bitte zu [[:$1]].',
-	'vet-warn1' => 'Sie müssen zuerst einen Suchbegriff eingeben!',
-	'vet-warn2' => 'Bitte geben Sie die URL des Videos an, das Sie einfügen möchten',
-	'vet-warn3' => 'Bitte geben Sie dem hinzuzufügenden Video einen Namen',
-	'vet-uploadtext' => "'''Dieses Formular erlaubt Ihnen das schnelle Hochladen und Einbinden von Dateien in Artikel.'''
-
-''Siehe auch das [[Special:Log/upload|Datei-Logbuch]] und die [[Special:ImageList|Liste aller hochgeladenen Dateien]].''",
-	'vet-notallowed' => 'Sie können in diesem Wiki keine Videos einbinden',
-	'vet-title' => 'Fügen Sie ein Video in den Artikel ein',
-	'vet-main-info' => '<i>Geben Sie die vollständige URL von einer der unterstützten Websites ein</i>',
-	'vet-page-success' => 'Die Video-Seite wurde aktualisiert. Sie müssen ggf. den Server-Cache per purge leeren um die Änderung zu sehen.',
-	'vet-no-preview' => 'Sie können aus dem Vorschaumodus kein Video hinzufügen',
-);
-
-/** Zazaki (Zazaki)
- * @author Erdemaslancan
- */
-$messages['diq'] = array(
-	'vet-upload' => 'GIRE Pêro',
-	'vet-thiswiki' => 'E na wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Bahdoyên',
-	'vet-find-btn' => 'Cı geyre',
-	'vet-name' => 'Name',
-	'vet-license' => 'Lisans',
-	'vet-size' => 'Gırdiye',
-	'vet-thumbnail' => 'Resmo qıckek',
-	'vet-fullsize' => 'Full-gırdiye ($1x$2px)',
-	'vet-width' => 'Herayey',
-	'vet-layout' => 'Ca gınayiş',
-	'vet-caption' => 'Bınnuşte',
-	'vet-optional' => '(Waştışo keyfi)',
-	'vet-prev' => "8'ê verêni",
-	'vet-next' => "8'ê peyêni",
-	'vet-return' => 'Peyd bê vurnayış',
-	'vet-back' => 'Peyd bê',
-	'vet-close' => 'Racnê',
-	'vet-show-message' => 'mesaci bımocne',
-	'vet-hide-message' => 'mesaci bınımne',
-	'vet-hide-license-msg' => 'nuşteya lisansi bınımne',
-	'vet-show-license-msg' => 'nuşteyê lisansi bımocne',
-	'vet-add-button' => 'Video deke',
-	'vet-see-all' => 'pêro',
-	'vet-left' => 'Çep',
-	'vet-right' => 'Raşt',
-	'vet-gallery' => 'Galeri',
-);
-
-/** Ewe (eʋegbe) */
-$messages['ee'] = array(
-	'vet-upload-btn' => 'Yi',
-	'vet-find-btn' => 'Dii',
-);
-
-/** Spanish (español)
- * @author Armando-Martin
- * @author McDutchie
- * @author Peter17
- * @author Sanbec
- * @author Translationista
- * @author VegaDark
- */
-$messages['es'] = array(
-	'vet-upload' => 'URL completa',
-	'vet-notlogged' => 'Identifícate o regístrate para insertar vídeos en este wiki',
-	'vet-find' => 'Buscar un vídeo',
-	'vet-thiswiki' => 'Este wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Siguiente',
-	'vet-find-btn' => 'Búsqueda',
-	'vet-flickr-inf' => 'Busca fotos con la licencia libre apropiada desde Flickr e impórtalas después a tu wiki.',
-	'vet-conflict-inf' => '<h1>Uups, ya existe un vídeo con ese nombre.</h1>¿Qué quieres hacer con <b>$1</b>?',
-	'vet-rename' => 'Renombrar Vídeo',
-	'vet-existing' => 'Usar el vídeo existente',
-	'vet-overwrite' => 'Sobrescribir el vídeo existente con el tuyo',
-	'vet-insert' => 'Insertar',
-	'vet-details-inf' => '<h1>Detalles del archivo</h1>Especifica un nombre para el vídeo. ¡Que sea descriptivo!',
-	'vet-name' => 'Nombre',
-	'vet-license' => 'Licencia',
-	'vet-details-inf2' => 'Opciones de Muestra del Vídeo',
-	'vet-size' => 'Tamaño',
-	'vet-thumbnail' => 'Miniatura',
-	'vet-fullsize' => 'Tamaño completo ($1x$2px)',
-	'vet-width' => 'Ancho',
-	'vet-layout' => 'Mostrar',
-	'vet-caption' => 'Comentario',
-	'vet-optional' => '(Opcional)',
-	'vet-insert2' => 'Insertar Vídeo',
-	'vet-update' => 'Actualizar video',
-	'vet-video-replace-link' => 'Para sustituir este video, por favor ve a [[:$1]]',
-	'vet-prev' => '8 anteriores',
-	'vet-next' => '8 siguientes',
-	'vet-metacafe' => 'Vídeos de Metacafe ({{PLURAL:$1|un resultado|$1 resultados}})',
-	'vet-insert3' => 'Insertar este vídeo',
-	'vet-recent-inf' => 'Vídeos Recientemente Insertados',
-	'vet-thiswiki2' => 'Vídeos de este Wiki ({{PLURAL:$1|un resultado|$1 resultados}})',
-	'vet-success' => '<h1>Éxito</h1>',
-	'vet-return' => 'Volver a la edición del artículo',
-	'vet-back' => 'volver',
-	'vet-close' => 'cerrar',
-	'vet-warn1' => '¡Necesitas especificar una pregunta antes!',
-	'vet-warn2' => 'Por favor, danos la url del vídeo que quieres añadir',
-	'vet-warn3' => 'Por favor, danos el nombre del vídeo que quieres añadir',
-	'vet-imagelink' => '<b>¡Nuevo!</b> Insertar Vídeos',
-	'vet-imagebutton' => 'Insertar Vídeos',
-	'vet-license-cc' => 'Licenciada bajo Creative Commons Attribution-Share Alike 3.0',
-	'vet-name-incorrect' => 'Este nombre de vídeo contiene caracteres inválidos, como #',
-	'vet-description' => 'Escribe la URL completa de cualquiera de los sitios soportados. Ver',
-	'vet-uploadtext' => "'''Este formulario te permite subir rápidamente y mostrar archivos en los artículos.''' ''Echa un vistazo también al [[Special:Log/upload|registro de subidas]] y la [[Special:ImageList|lista de todos los archivos subidos]].''",
-	'vet-show-message' => 'mostrar mensaje',
-	'vet-hide-message' => 'ocultar mensaje',
-	'vet-hide-license-msg' => 'ocultar texto de la licencia',
-	'vet-show-license-msg' => 'mostrar el texto de la licencia',
-	'vet-max-thumb' => 'tamaño máximo del thumbnail excedido, revirtiendo a su tamaño original',
-	'vet-notallowed' => 'No puedes insertar vídeos en este wiki',
-	'vet-title' => 'Añadir un vídeo a tu artículo',
-	'vet-supported' => 'Sitios de vídeos soportados:',
-	'vet-preview' => '<i>esta es solamente una vista reducida de lo que se mostrará</i>',
-	'vet-bad-url' => 'La URL aportada es inválida',
-	'vet-non-existing' => 'El vídeo proporcionado no existe',
-	'vet-main-info' => '<i>Pon la URL completa del vídeo contenido en uno de los sitios soportados que se muestran debajo.</i>',
-	'vet-page-success' => 'La página del vídeo fue actualizada satisfactoriamente. Debes recargar la página del vídeo para ver los cambios.',
-	'vet-add-button' => 'Añadir video',
-	'vet-added-from-gallery' => 'Video agregado desde la galería de video',
-	'vet-see-all' => 'ver todos',
-	'vet-bad-search' => 'Error al cargar la consulta de búsqueda.',
-	'vet-no-preview' => 'No puedes añadir un vídeo a un artículo desde el modo de previsualización',
-	'vet-single-success' => 'El video ha sido colocado con éxito en este artículo',
-	'vet-gallery-create-success' => 'Se ha añadido una galería de video a este artículo',
-	'vet-gallery-add-success' => 'Se ha añadido el video a la galería de video en este artículo',
-	'vet-left' => 'Izquierda',
-	'vet-right' => 'Derecha',
-	'vet-gallery' => 'Galería',
-	'vet-protected' => 'Este video está protegido',
-);
-
-/** Basque (euskara)
- * @author An13sa
- */
-$messages['eu'] = array(
-	'vet-find' => 'Bideoa bilatu',
-	'vet-flickr' => 'Flickr',
-	'vet-find-btn' => 'Bilaketa',
-	'vet-insert' => 'Txertatu',
-	'vet-name' => 'Izena',
-	'vet-license' => 'Lizentzia',
-);
-
-/** Finnish (suomi)
- * @author Crt
- * @author Silvonen
- * @author Tofu II
- */
-$messages['fi'] = array(
-	'vet-find' => 'Etsi video',
-	'vet-thiswiki' => 'Tämä wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Seuraava',
-	'vet-name' => 'Nimi',
-	'vet-size' => 'Koko',
-	'vet-thumbnail' => 'Pienoiskuva',
-	'vet-width' => 'Leveys',
-	'vet-update' => 'Päivitä video',
-	'vet-back' => 'takaisin',
-	'vet-close' => 'sulje',
-	'vet-show-message' => 'näytä viesti',
-	'vet-hide-message' => 'piilota viesti',
-	'vet-hide-license-msg' => 'piilota lisenssiteksti',
-	'vet-show-license-msg' => 'näytä lisenssiteksti',
-	'vet-title' => 'Lisää video artikkeliin',
-	'vet-add-button' => 'Lisää video',
-	'vet-gallery' => 'Galleria',
-	'vet-protected' => 'Tämä video on suojattu',
-);
-
-/** French (français)
- * @author Crochet.david
- * @author Dodoïste
- * @author Gomoko
- * @author IAlex
- * @author McDutchie
- * @author Peter17
- * @author Slamduck
- * @author Wyz
- */
-$messages['fr'] = array(
-	'vet-upload' => 'URL complète',
-	'vet-notlogged' => "Connectez-vous ou créez un compte d'utilisateur pour incorporer des vidéos",
-	'vet-find' => 'Chercher une vidéo',
-	'vet-thiswiki' => 'Ce wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Suivant',
-	'vet-find-btn' => 'Chercher',
-	'vet-flickr-inf' => 'Rechercher des photos de licence libre à partir de Flickr et les incorporer dans votre wiki.',
-	'vet-conflict-inf' => "<h1>Il y a déjà une vidéo existant avec ce nom !</h1>Qu'est-ce que vous voulez faire avec <b>$1</b> ?",
-	'vet-rename' => 'Renommer la vidéo',
-	'vet-existing' => 'Utiliser la vidéo existante',
-	'vet-overwrite' => 'Remplacer la vidéo existante par votre vidéo.',
-	'vet-insert' => 'Insérer',
-	'vet-details-inf' => '<h1>Le nom du fichier</h1>Merci de donner un nom très significatif pour la vidéo.',
-	'vet-name' => 'Nom',
-	'vet-license' => 'Licence',
-	'vet-details-inf2' => 'Options d’affichage vidéo',
-	'vet-size' => 'Taille',
-	'vet-thumbnail' => 'Aperçu',
-	'vet-fullsize' => 'Taille originale ($1px/$2px)',
-	'vet-width' => 'Largeur',
-	'vet-layout' => 'Orientation',
-	'vet-caption' => 'Description',
-	'vet-optional' => '(Optionnel)',
-	'vet-insert2' => 'Incorporer',
-	'vet-update' => 'Mettre à jour la vidéo',
-	'vet-video-replace-link' => 'Pour remplacer cette vidéo, veuillez vous diriger vers [[:$1]]',
-	'vet-prev' => '8 précédentes',
-	'vet-next' => '8 suivantes',
-	'vet-metacafe' => 'Vidéos de Metacafe ($1 {{PLURAL:$1|résultat|résultats}})',
-	'vet-insert3' => 'Incorper cette vidéo',
-	'vet-recent-inf' => 'Vidéos récemment ajoutées',
-	'vet-thiswiki2' => 'Vidéos sur ce wiki ($1 {{PLURAL:$1|résultat|résultats}})',
-	'vet-success' => '<h1>Succès</h1>',
-	'vet-return' => 'Retourner à la modification',
-	'vet-back' => 'Retour',
-	'vet-close' => 'fermer',
-	'vet-warn1' => "Vous devez d'abord spécifier votre requête!",
-	'vet-warn2' => "Merci d'ajouter l'URL de la vidéo que vous voulez importer sur {{SITENAME}}",
-	'vet-warn3' => 'Merci de donner un nom à la vidéo que vous importez.',
-	'vet-imagelink' => '<b>Nouveau!</b> Ajouter des vidéos',
-	'vet-imagebutton' => 'Ajouter des vidéos',
-	'vet-license-cc' => 'Disponible sous licence Creative Commons Attribution-Share Alike 3.0',
-	'vet-name-incorrect' => 'Le nom de cette vidéo contient des caractères non-valides, comme #',
-	'vet-description' => "Entrez l'URL complète provenant de l'un des sites pris en charge. Voir",
-	'vet-uploadtext' => "'''Ce formulaire vous permet d'importer rapidement des fichiers et de les afficher dans les articles.'''
-
-''Allez voir aussi l'[[Special:Log/upload|historique des imports de fichiers multimédia]] et la [[Special:ImageList|liste de tout les fichiers importés]].''",
-	'vet-show-message' => 'Afficher le message',
-	'vet-hide-message' => 'masquer le message',
-	'vet-hide-license-msg' => 'Masquer le texte de licence',
-	'vet-show-license-msg' => 'Afficher la licence',
-	'vet-max-thumb' => "la taille maximale de l'image a été atteinte, retour à la taille originale",
-	'vet-notallowed' => 'Vous ne pouvez pas importer de vidéos sur ce wiki',
-	'vet-title' => "Ajouter une vidéo dans l'article",
-	'vet-supported' => 'Sites de vidéos pris en charge :',
-	'vet-preview' => "''Cet aperçu n'est pas à la taille réelle''",
-	'vet-bad-url' => "Cette adresse URL n'est pas valide. Vérifiez l'adresse, s'il vous plaît.",
-	'vet-non-existing' => "La vidéo ajoutée n'existe pas",
-	'vet-main-info' => "<i>Entrez l'URL complète provenant de l'un des sites pris en charge.</i>",
-	'vet-page-success' => 'La page de la vidéo a été mise à jour avec succès. Vous devez purger la page de la vidéo pour voir les changements.',
-	'vet-add-button' => 'Ajouter une vidéo',
-	'vet-added-from-gallery' => "Vidéo ajoutée d'une galerie de vidéos",
-	'vet-see-all' => 'tous',
-	'vet-bad-search' => 'Une erreur est survenue durant le chargement de la requête',
-	'vet-no-preview' => 'Vous ne pouvez pas ajouter de vidéos lors du mode de prévisualisation',
-	'vet-single-success' => 'La vidéo a été placée avec succès dans l’article.',
-	'vet-gallery-create-success' => "Une galerie de vidéos fut ajoutée à l'article",
-	'vet-gallery-add-success' => 'Une vidéo a été ajoutée dans la galerie de vidéos de cet article',
-	'vet-left' => 'Gauche',
-	'vet-right' => 'Droite',
-	'vet-gallery' => 'Galerie',
-	'vet-protected' => 'Cette vidéo est protégée',
-);
-
-/** Galician (Galego)
- * @author Toliño
- */
-$messages['gl'] = array(
-	'vet-upload' => 'URL completo',
-	'vet-notlogged' => 'Acceda ao sistema ou rexístrese para inserir vídeos',
-	'vet-find' => 'Atopar un vídeo',
-	'vet-thiswiki' => 'Este wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Seguinte',
-	'vet-find-btn' => 'Procurar',
-	'vet-flickr-inf' => 'Busca fotografías de licenza libre de Flickr e impórtaas no seu wiki.',
-	'vet-conflict-inf' => '<h1>Vaites!, xa existe un vídeo con ese nome.</h1>Que quere facer con <b>$1</b>?',
-	'vet-rename' => 'Renomear o vídeo',
-	'vet-existing' => 'Usar o vídeo existente',
-	'vet-overwrite' => 'Substituír o vídeo existente co seu',
-	'vet-insert' => 'Inserir',
-	'vet-details-inf' => '<h1>Detalles do ficheiro</h1>Especifique un nome para este vídeo. Escolla un que sexa descritivo.',
-	'vet-name' => 'Nome',
-	'vet-license' => 'Licenza',
-	'vet-details-inf2' => 'Opcións de visualización do vídeo',
-	'vet-size' => 'Tamaño',
-	'vet-thumbnail' => 'Miniatura',
-	'vet-fullsize' => 'Tamaño completo ($1x$2px)',
-	'vet-width' => 'Ancho',
-	'vet-layout' => 'Disposición',
-	'vet-caption' => 'Pé de foto',
-	'vet-optional' => '  (opcional)',
-	'vet-insert2' => 'Inserir o vídeo',
-	'vet-update' => 'Actualizar o vídeo',
-	'vet-video-replace-link' => 'Para substituír este vídeo, vaia a [[:$1]]',
-	'vet-prev' => '8 anteriores',
-	'vet-next' => '8 seguintes',
-	'vet-metacafe' => 'Vídeos de Metacafe ($1  {{PLURAL:$1|resultado|resultados}})',
-	'vet-insert3' => 'Inserir este vídeo',
-	'vet-recent-inf' => 'Vídeos inseridos recentemente',
-	'vet-thiswiki2' => 'Vídeos deste wiki ($1 {{PLURAL:$1|resultado|resultados}})',
-	'vet-success' => '<h1>Feito</h1>',
-	'vet-return' => 'Volver ao modo de edición',
-	'vet-back' => 'Volver',
-	'vet-close' => 'Pechar',
-	'vet-warn1' => 'Ten que especificar a pescuda primeiro!',
-	'vet-warn2' => 'Proporcione o enderezo URL do vídeo que lle gustaría engadir',
-	'vet-warn3' => 'Proporcione o nome do vídeo que lle gustaría engadir',
-	'vet-imagelink' => '<b>Novo!</b> Inserir vídeos',
-	'vet-imagebutton' => 'Inserir vídeos',
-	'vet-license-cc' => 'Licenciado baixo a licenza Creative Commons recoñecemento compartir igual 3.0',
-	'vet-name-incorrect' => 'O nome deste vídeo contén caracteres non válidos, como #',
-	'vet-description' => 'Escriba o enderezo URL completo de calquera dos sitios soportados. Véxase',
-	'vet-uploadtext' => "'''Este formulario permítelle cargar ficheiros rapidamente e incluílos en artigos.'''
-
-''Olle tamén o [[Special:Log/upload|rexistro de cargas]] e mais unha [[Special:ImageList|lista con todos os ficheiros cargados]].''",
-	'vet-show-message' => 'mostrar a mensaxe',
-	'vet-hide-message' => 'agochar a mensaxe',
-	'vet-hide-license-msg' => 'agochar o texto da licenza',
-	'vet-show-license-msg' => 'mostrar o texto da licenza',
-	'vet-max-thumb' => 'superouse o tamaño máximo para unha miniatura; volvendo ao tamaño orixinal',
-	'vet-notallowed' => 'Non pode inserir vídeos neste wiki',
-	'vet-title' => 'Engadir un vídeo ao seu artigo',
-	'vet-supported' => 'Sitios de vídeos soportados:',
-	'vet-preview' => '<i>Esta vista previa non está a escala</i>',
-	'vet-bad-url' => 'O enderezo URL solicitado non é válido',
-	'vet-non-existing' => 'O vídeo solicitado non existe',
-	'vet-main-info' => '<i>Escriba o enderezo URL completo de calquera dos sitios soportados</i>',
-	'vet-page-success' => 'A páxina do vídeo foi actualizada correctamente. Poida que teña que recargar a páxina para ollar os cambios.',
-	'vet-add-button' => 'Engadir o vídeo',
-	'vet-added-from-gallery' => 'vídeo engadido dunha galería de vídeos',
-	'vet-see-all' => 'todos',
-	'vet-bad-search' => 'Erro ao cargar a pescuda.',
-	'vet-no-preview' => 'Non pode engadir un vídeo ao artigo no modo de vista previa',
-	'vet-single-success' => 'O vídeo colocouse con éxito neste artigo',
-	'vet-gallery-create-success' => 'Engadiuse unha galería de vídeos a este artigo',
-	'vet-gallery-add-success' => 'O vídeo engadiuse á galería de vídeos deste artigo',
-	'vet-left' => 'Esquerda',
-	'vet-right' => 'Dereita',
-	'vet-gallery' => 'Galería',
-	'vet-protected' => 'Este vídeo está protexido',
-);
-
-/** Hausa (هَوُسَ) */
-$messages['ha'] = array(
-	'vet-find-btn' => 'Nema',
-);
-
-/** Hebrew (עברית)
- * @author שומבלע
- */
-$messages['he'] = array(
-	'vet-upload' => 'כתובת מלאה',
-	'vet-notlogged' => 'התחברו או הרשמו כדי להוסיף סרטונים',
-	'vet-find' => 'חיפוש סרטון',
-	'vet-thiswiki' => 'באתר',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'המשך',
-	'vet-find-btn' => 'חיפוש',
-	'vet-flickr-inf' => 'חיפוש תמונות ברישיון חופשי ב־Flickr והוספה לאתר.',
-	'vet-conflict-inf' => '<h1 style=";text-align:right;direction:rtl"> אופס, סרטון עם שם זהה כבר קיים. </h1> מה ברצונך לעשות עם <b>$1</b>?',
-	'vet-rename' => 'שינוי שם הסרטון',
-	'vet-existing' => 'שימוש בסרטונים קיימים',
-	'vet-overwrite' => 'החלפת הסרטון הקיים באחר',
-	'vet-insert' => 'הוספה',
-	'vet-details-inf' => '<h1>פרטי הקובץ</h1>בחירת שם לסרטון.השם צריך לתאר אותו.',
-	'vet-name' => 'שם',
-	'vet-license' => 'רישיון',
-	'vet-details-inf2' => 'אפשרויות תצוגה',
-	'vet-size' => 'גודל',
-	'vet-thumbnail' => 'ממוזער',
-	'vet-fullsize' => 'גודל מלא ($1x$2px)',
-	'vet-width' => 'רוחב',
-	'vet-layout' => 'פריסה',
-	'vet-caption' => 'תיאור',
-	'vet-optional' => '(אפשרי)',
-	'vet-insert2' => 'הוספת סרטון',
-	'vet-update' => 'עדכון סרטון',
-	'vet-video-replace-link' => 'כדי להחליף את הסרטון, נא לעבור אל [[:$1]]',
-	'vet-prev' => '8 הקודמים',
-	'vet-next' => '8 הבאים',
-	'vet-metacafe' => 'סרטוני Metacafe ($1 {{PLURAL:$1|תוצאה|תוצאות}})',
-	'vet-insert3' => 'הוספת סרטון',
-	'vet-recent-inf' => 'סרטונים שנוספו לאחרונה',
-	'vet-thiswiki2' => 'סרטונים מהאתר ($1 {{PLURAL:$1|תוצאה|תוצאות}})',
-	'vet-success' => '<h1 style=";text-align:right;direction:rtl"> הצלחה </h1>',
-	'vet-return' => 'חזרה לעריכה',
-	'vet-back' => 'חזרה',
-	'vet-close' => 'סגירה',
-	'vet-warn1' => 'צריך לציין את שם הקובץ קודם!',
-	'vet-warn2' => 'נא לצרף כתובת הסרטון להוספה',
-	'vet-warn3' => 'נא לכתוב את שם הסרטון להוספה',
-	'vet-imagelink' => '<b>חדש!</b> הוספת סרטונים',
-	'vet-imagebutton' => 'הוספת סרטונים',
-	'vet-license-cc' => 'רשיון לפי Creative Commons Attribution Share-Alike 3.0',
-	'vet-name-incorrect' => 'שם הסרטון מכיל תווים לא חוקיים',
-	'vet-uploadtext' => "'''טופס זה מאפשר לך להעלות ולהציג קבצי הדפים בקלות.''' 
-
-''ראו גם  [[Special:Log/upload|יומן העלאות]] ו[[Special:ImageList|רשימת כל הקבצים באתר]].''",
-	'vet-show-message' => 'הצגת הודעה',
-	'vet-hide-message' => 'הסתרת הודעה',
-	'vet-hide-license-msg' => 'הסתרת הרישיון',
-	'vet-show-license-msg' => 'הצגת הרישיון',
-	'vet-max-thumb' => 'חריגה מגודל סרטון מקסימלי ,מחזיר לגודל המקורי',
-	'vet-notallowed' => 'לא ניתן להוסיף סרטונים לאתר.',
-	'vet-title' => 'הוספת סרטון לדף',
-	'vet-supported' => 'אתרי וידאו נתמכים:',
-	'vet-preview' => '<i>תצוגה זו לא בקנה מידה</i>',
-	'vet-bad-url' => 'כתובת האתר שצורפה אינה חוקית',
-	'vet-non-existing' => 'הסרטון  שצורף אינו קיים',
-	'vet-main-info' => '<i>חיפוש סרטונים או הכנסת כתובת הסרטון המלאה מאחד האתרים הנתמכים</i>',
-	'vet-page-success' => 'דף הסרטונים עודכן בהצלחה. ייתכן שיהיה צורך לרענן את דף הסרטונים כדי לראות את השינוי.',
-	'vet-add-button' => 'הוסף סרטון',
-	'vet-added-from-gallery' => 'סרטון נוסף מתוך גלריית הסרטונים',
-	'vet-see-all' => 'הצג הכל',
-	'vet-bad-search' => 'שגיאה בטעינת שאילתת החיפוש.',
-	'vet-no-preview' => 'לא ניתן להוסיף סרטון מדף במצב תצוגה מקדימה',
-	'vet-single-success' => 'הסרטון הוצב בדף בהצלחה.',
-	'vet-gallery-create-success' => 'גלריית הסרטונים נוספה לדף.',
-	'vet-gallery-add-success' => 'הסרטון נוסף לגלריית הסרטונים בדף.',
-	'vet-left' => 'שמאל',
-	'vet-right' => 'ימין',
-	'vet-gallery' => 'גלריה',
-	'vet-protected' => 'סרטון זה מוגן.',
-);
-
-/** Croatian (hrvatski)
- * @author Anton008
- */
-$messages['hr'] = array(
-	'vet-left' => 'Lijevo',
-);
-
-/** Hungarian (magyar)
- * @author Glanthor Reviol
- */
-$messages['hu'] = array(
-	'vet-upload' => 'Teljes URL',
-	'vet-find' => 'Videó keresése',
-	'vet-thiswiki' => 'Ez a Wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Feltöltés',
-	'vet-find-btn' => 'Keresés',
-	'vet-rename' => 'Videó átnevezése',
-	'vet-existing' => 'Létező videó használata',
-	'vet-overwrite' => 'Jelenlegi videó felülírása a tiéddel',
-	'vet-insert' => 'Beillesztés',
-	'vet-name' => 'Név',
-	'vet-license' => 'Licenc',
-	'vet-details-inf2' => 'Videó megjelenítési beállítások',
-	'vet-size' => 'Méret',
-	'vet-thumbnail' => 'Bélyegkép',
-	'vet-fullsize' => 'Teljes méret ($1×$2 képpont)',
-	'vet-width' => 'Szélesség',
-	'vet-optional' => ' (nem kötelező)',
-	'vet-insert2' => 'Videó beillesztése',
-	'vet-update' => 'Videó frissítése',
-	'vet-prev' => 'Előző 8',
-	'vet-next' => 'Következő 8',
-	'vet-insert3' => 'Videó beillesztése',
-	'vet-success' => '<h1>Siker</h>',
-	'vet-return' => 'Vissza a szerkesztéshez',
-	'vet-back' => 'vissza',
-	'vet-close' => 'bezárás',
-	'vet-imagelink' => '<b>Újdonság!</b> Videók beszúrása',
-	'vet-imagebutton' => 'Videók beillesztése',
-	'vet-show-message' => 'üzenet megjelenítése',
-	'vet-hide-message' => 'üzenet elrejtése',
-	'vet-hide-license-msg' => 'licenc szövegének elrejtése',
-	'vet-show-license-msg' => 'licenc szövegének megjelenítése',
-	'vet-supported' => 'Támogatott videomegosztó oldalak:',
-	'vet-bad-url' => 'A megadott URL érvénytelen',
-	'vet-add-button' => 'Videó hozzáadása',
-	'vet-see-all' => 'összes megjelenítése',
-	'vet-bad-search' => 'Hiba a keresési lekérdezés betöltésekor.',
-	'vet-left' => 'Bal',
-	'vet-right' => 'Jobb',
-	'vet-gallery' => 'Galéria',
-	'vet-protected' => 'Ez a videó védett',
-);
-
-/** Interlingua (interlingua)
- * @author McDutchie
- */
-$messages['ia'] = array(
-	'vet-upload' => 'URL complete',
-	'vet-notlogged' => 'Aperi un session o crea un conto pro inserer videos',
-	'vet-find' => 'Cercar un video',
-	'vet-thiswiki' => 'Iste wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Sequente',
-	'vet-find-btn' => 'Cercar',
-	'vet-flickr-inf' => 'Cerca photos de licentia libere in Flickr e importa los in tu wiki.',
-	'vet-conflict-inf' => '<h1>Ups, un video con iste nomine ja existe.</h1>Que vole tu facer con <b>$1</b>?',
-	'vet-rename' => 'Renominar video',
-	'vet-existing' => 'Usar video existente',
-	'vet-overwrite' => 'Reimplaciar le video existente per le tue',
-	'vet-insert' => 'Inserer',
-	'vet-details-inf' => '<h1>Detalios del file</h1>Specifica un nomine descriptive pro iste video.',
-	'vet-name' => 'Nomine',
-	'vet-license' => 'Licentia',
-	'vet-details-inf2' => 'Optiones de reproduction video',
-	'vet-size' => 'Dimension',
-	'vet-thumbnail' => 'Miniatura',
-	'vet-fullsize' => 'Dimension original ($1 × $2 px)',
-	'vet-width' => 'Latitude',
-	'vet-layout' => 'Orientation',
-	'vet-caption' => 'Legenda',
-	'vet-optional' => '  (Optional)',
-	'vet-insert2' => 'Inserer video',
-	'vet-update' => 'Actualisar video',
-	'vet-video-replace-link' => 'Pro reimplaciar iste video, per favor visita [[:$1]]',
-	'vet-prev' => '8 precedentes',
-	'vet-next' => '8 sequentes',
-	'vet-metacafe' => 'Videos de Metacafe ($1 {{PLURAL:$1|resultato|resultatos}})',
-	'vet-insert3' => 'Inserer iste video',
-	'vet-recent-inf' => 'Videos recentemente inserite',
-	'vet-thiswiki2' => 'Videos in iste wiki ($1 {{PLURAL:$1|resultato|resultatos}})',
-	'vet-success' => '<h1>Successo</h1>',
-	'vet-return' => 'Retornar al modification',
-	'vet-back' => 'retornar',
-	'vet-close' => 'clauder',
-	'vet-warn1' => 'Tu debe primo specificar le texto a cercar!',
-	'vet-warn2' => 'Per favor forni le URL del video que tu vole adder',
-	'vet-warn3' => 'Per favor forni le nomine del video que tu vole adder',
-	'vet-imagelink' => '<b>Nove!</b> Insertion de videos',
-	'vet-imagebutton' => 'Inserer videos',
-	'vet-license-cc' => 'Disponibile sub licentia Creative Commons Attribution-Share Alike 3.0',
-	'vet-name-incorrect' => 'Le nomine de iste video contine characteres invalide, como #',
-	'vet-description' => 'Entra le URL complete de un del sitos supportate. Vide',
-	'vet-uploadtext' => "'''Iste formulario permitte rapidemente incargar e presentar files in articulos.'''
-
-''Vide etiam le [[Special:Log/upload|registro de incargamentos]] e un [[Special:ImageList|lista de tote le files incargate]].''",
-	'vet-show-message' => 'revelar message',
-	'vet-hide-message' => 'celar message',
-	'vet-hide-license-msg' => 'celar texto del licentia',
-	'vet-show-license-msg' => 'revelar texto del licentia',
-	'vet-max-thumb' => 'dimension maximal del miniatura excedite, reversion al dimension original',
-	'vet-notallowed' => 'Tu non pote inserer videos in iste wiki.',
-	'vet-title' => 'Adder un video a tu articulo',
-	'vet-supported' => 'Sitos de video supportate:',
-	'vet-preview' => '<i>Previsualisation a scala reducite</i>',
-	'vet-bad-url' => 'Le URL fornite es invalide',
-	'vet-non-existing' => 'Le video fornite non existe',
-	'vet-main-info' => '<i>Entra le URL complete de un del sitos supportate</i>',
-	'vet-page-success' => 'Le pagina del video ha essite actualisate. Es possibile que tu debe recargar le pagina del video pro vider le cambio.',
-	'vet-add-button' => 'Adder video',
-	'vet-added-from-gallery' => 'video addite ex galeria de videos',
-	'vet-see-all' => 'totes',
-	'vet-bad-search' => 'Error durante le cargamento del consulta de recerca.',
-	'vet-no-preview' => 'Tu non pote adder un video ex un articulo in modo de previsualisation.',
-	'vet-single-success' => 'Le video ha essite placiate in iste articulo con successo.',
-	'vet-gallery-create-success' => 'Un galeria de videos ha essite addite a iste articulo.',
-	'vet-gallery-add-success' => 'Un video ha essite addite al galeria de videos in iste articulo.',
-	'vet-left' => 'Sinistra',
-	'vet-right' => 'Dextra',
-	'vet-gallery' => 'Galeria',
-	'vet-protected' => 'Iste video es protegite',
-);
-
-/** Indonesian (Bahasa Indonesia)
- * @author Bennylin
- * @author Irwangatot
- * @author Kenrick95
- */
-$messages['id'] = array(
-	'vet-insert' => 'Sisipkan',
-	'vet-name' => 'Nama',
-	'vet-license' => 'Lisensi',
-	'vet-details-inf2' => 'Pilihan Tampilan Video',
-	'vet-size' => 'Ukuran',
-	'vet-thumbnail' => 'Miniatur',
-	'vet-fullsize' => 'Ukuran-penuh ($1x$2px)',
-	'vet-width' => 'Lebar',
-	'vet-layout' => 'Tampilan',
-	'vet-caption' => 'Keterangan',
-	'vet-optional' => '(Pilihan)',
-	'vet-insert2' => 'Masukkan Video',
-	'vet-update' => 'Perbaharui Video',
-	'vet-video-replace-link' => 'Untuk mengganti video ini, silakan ke [[:$1]]',
-	'vet-prev' => '8 Sebelumnya',
-	'vet-next' => '8 Berikutnya',
-	'vet-metacafe' => 'Metacafe Videos ($1 hasil)',
-	'vet-insert3' => 'Masukkan video ini',
-	'vet-recent-inf' => 'Video terbaru dimasukan',
-	'vet-thiswiki2' => 'Video Wiki ini ($1 hasil)',
-	'vet-success' => '<h1>Berhasil</h1>',
-	'vet-return' => 'Kembali menyunting',
-	'vet-back' => 'kembali',
-	'vet-close' => 'tutup',
-	'vet-warn1' => 'Anda perlu menentukan permintaan terlebih dahulu!',
-	'vet-warn2' => 'Harap berikan URL dari video yang ingin anda ditambahkan',
-	'vet-warn3' => 'Harap berikan nama video yang ingin anda menambahkan',
-	'vet-imagelink' => '<b>Baru!</b> Sisipkan Video',
-	'vet-imagebutton' => 'Sisipkan Video',
-	'vet-license-cc' => 'dilisensikan di bawah Creative Commons Attribution-Share Alike 3,0',
-	'vet-name-incorrect' => 'Nama video ini berisi karakter tidak sah',
-	'vet-show-message' => 'tampilkan pesan',
-	'vet-hide-message' => 'sembunyikan pesan',
-	'vet-hide-license-msg' => 'sembunyikan teks lisensi',
-	'vet-show-license-msg' => 'tampilkan teks lisensi',
-	'vet-max-thumb' => 'ukuran thumbnail maksimum terlampaui, kembali ke ukuran aslinya',
-	'vet-notallowed' => 'Anda tidak dapat menyisipkan video di wiki ini',
-	'vet-title' => 'Tambahkan video ke artikel Anda',
-	'vet-supported' => 'Situs video yang didukung:',
-	'vet-preview' => '<i>Pratayang ini bukan untuk skala</i>',
-	'vet-bad-url' => 'URL yang diberikan tidak syah',
-	'vet-non-existing' => 'Video yang disertakan tidak ada',
-	'vet-main-info' => '<i>Mencari video atau memasukkan URL lengkap dari salah satu situs yang didukung</i>',
-	'vet-add-button' => 'Tambah Video',
-	'vet-added-from-gallery' => 'menambah video dari galeri video',
-	'vet-see-all' => 'lihat semua',
-	'vet-no-preview' => 'Anda tidak dapat menambahkan video dalam artikel dengan modus pratayang',
-	'vet-single-success' => 'Video berhasil disisipkan dalam artikel ini',
-	'vet-gallery-create-success' => 'Sebuah galeri video telah ditambahkan ke artikel ini',
-	'vet-gallery-add-success' => 'Video telah ditambahkan ke galeri video dalam artikel ini',
-	'vet-left' => 'Kiri',
-	'vet-right' => 'Kanan',
-	'vet-gallery' => 'Galeri',
-	'vet-protected' => 'Video ini dilindungi',
-);
-
-/** Igbo (Igbo)
- * @author Ukabia
- */
-$messages['ig'] = array(
-	'vet-find-btn' => 'Chọwa',
-);
-
-/** Ingush (ГӀалгӀай)
- * @author Sapral Mikail
- */
-$messages['inh'] = array(
-	'vet-gallery' => 'Суртанче',
-);
-
-/** Italian (italiano)
- * @author Beta16
- * @author HalphaZ
- */
-$messages['it'] = array(
-	'vet-find' => 'Trova un video',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'VAI',
-	'vet-find-btn' => 'Ricerca',
-	'vet-flickr-inf' => 'Cerca foto da Flickr con regolare licenza libera e importale nel tuo wiki.',
-	'vet-rename' => 'Rinomina video',
-	'vet-existing' => 'Usa un video esistente',
-	'vet-insert' => 'Inserisci',
-	'vet-name' => 'Nome',
-	'vet-license' => 'Licenza',
-	'vet-size' => 'Dimensione',
-	'vet-thumbnail' => 'Miniatura',
-	'vet-width' => 'Larghezza',
-	'vet-layout' => 'Formattazione',
-	'vet-caption' => 'Didascalia',
-	'vet-optional' => '(opzionale)',
-	'vet-insert2' => 'Inserisci video',
-	'vet-update' => 'Aggiorna video',
-	'vet-prev' => 'Precedenti 8',
-	'vet-next' => 'Succ 8',
-	'vet-insert3' => 'Inserisci questo video',
-	'vet-recent-inf' => 'Video inseriti recentemente',
-	'vet-back' => 'indietro',
-	'vet-close' => 'Chiudi',
-	'vet-imagebutton' => 'Inserisci video',
-	'vet-show-message' => 'mostra messaggio',
-	'vet-hide-message' => 'nascondi messaggio',
-	'vet-hide-license-msg' => 'Nascondi il testo della licenza',
-	'vet-show-license-msg' => 'Mostra il testo della licenza',
-	'vet-title' => 'Aggiungi un video al tuo articolo',
-	'vet-bad-url' => "L'URL fornito non è valido",
-	'vet-non-existing' => 'Il video fornito non esiste',
-	'vet-add-button' => 'Aggiungi video',
-	'vet-see-all' => 'vedi tutti',
-	'vet-left' => 'Sinistra',
-	'vet-right' => 'Giusto',
-	'vet-gallery' => 'Galleria',
-	'vet-protected' => 'Questo video è protetto',
-);
-
-/** Japanese (日本語)
- * @author Schu
- * @author Tommy6
- */
-$messages['ja'] = array(
-	'vet-upload' => 'URL',
-	'vet-notlogged' => '動画の挿入にはログインかアカウントの登録が必要です。',
-	'vet-find' => '動画を探す',
-	'vet-thiswiki' => 'このウィキ',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => '次',
-	'vet-find-btn' => '検索',
-	'vet-flickr-inf' => 'Flickrにある適切なライセンスのフリーな画像を検索し、ウィキにインポートします。',
-	'vet-conflict-inf' => '<h1>同じ名前の動画が既にあります。</h1><b>$1</b> に対する操作を選択してください。',
-	'vet-rename' => '動画名を変更',
-	'vet-existing' => '既にある動画を使用',
-	'vet-overwrite' => '既にある動画に上書き',
-	'vet-insert' => '挿入',
-	'vet-details-inf' => '<h1>ファイルの詳細</h1>この動画の名前を指定してください。動画の内容がわかるようなものがいいでしょう。',
-	'vet-name' => '動画名',
-	'vet-license' => 'ライセンス',
-	'vet-details-inf2' => '動画表示のオプション',
-	'vet-size' => 'サイズ',
-	'vet-thumbnail' => 'サムネイル',
-	'vet-fullsize' => 'フルサイズ ($1 x $2 ピクセル)',
-	'vet-width' => '幅',
-	'vet-layout' => 'レイアウト',
-	'vet-caption' => 'キャプション',
-	'vet-optional' => '（オプション）',
-	'vet-insert2' => '動画を挿入',
-	'vet-update' => '動画を更新',
-	'vet-video-replace-link' => 'この動画を上書きするには、[[:$1]]に行ってください。',
-	'vet-prev' => '前の8件',
-	'vet-next' => '次の8件',
-	'vet-metacafe' => 'Metacafe の動画 ($1 件)',
-	'vet-insert3' => 'この動画を挿入',
-	'vet-recent-inf' => '最近挿入された動画',
-	'vet-thiswiki2' => 'このウィキの動画 ($1 件)',
-	'vet-success' => '<h1>成功しました</h1>',
-	'vet-return' => '編集に戻る',
-	'vet-back' => '戻る',
-	'vet-close' => '閉じる',
-	'vet-warn1' => '検索する語を入力してください',
-	'vet-warn2' => '追加したい動画のURLを入力してください',
-	'vet-warn3' => '追加する動画の名前を入力してください',
-	'vet-imagelink' => '<b>New!</b> 動画挿入機能',
-	'vet-imagebutton' => '動画を挿入',
-	'vet-license-cc' => 'CC-by-sa 3.0を適用する',
-	'vet-name-incorrect' => '動画名に無効な文字が含まれています',
-	'vet-description' => 'サポートされているサイトのいずれかから完全なURLを入力してください。参照',
-	'vet-uploadtext' => "'''このフォームでは、簡単にファイルをアップロードして記事に挿入できます。'''
-
-''[[Special:Log/upload|アップロード記録]]や[[Special:ImageList|アップロードされたファイルのリスト]]もご覧ください。''",
-	'vet-show-message' => 'メッセージを表示',
-	'vet-hide-message' => 'メッセージを非表示',
-	'vet-hide-license-msg' => 'ライセンステキストを非表示',
-	'vet-show-license-msg' => 'ライセンステキストを表示',
-	'vet-max-thumb' => 'サムネイルのサイズが制限を越えるため、オリジナルのサイズに戻します',
-	'vet-notallowed' => 'このウィキでは動画を挿入できません',
-	'vet-title' => '記事に動画を追加',
-	'vet-supported' => '対応動画サイト:',
-	'vet-preview' => '縮小プレビュー',
-	'vet-bad-url' => '入力されたURLが不適切です。',
-	'vet-non-existing' => '指定された動画は存在しません',
-	'vet-main-info' => '<i>サポートされているサイトのいずれかから完全なURLを入力してください。</i>',
-	'vet-page-success' => '動画ページの更新に成功しました。変更を確認するには動画ページをパージする必要があります。',
-	'vet-add-button' => '動画を追加',
-	'vet-added-from-gallery' => '動画ギャラリーから動画を追加',
-	'vet-see-all' => 'すべて',
-	'vet-bad-search' => '検索クエリのロードでエラーが発生しました。',
-	'vet-no-preview' => 'プレビューモードでは記事に動画を追加できません',
-	'vet-single-success' => '動画の記事中への挿入に成功しました',
-	'vet-gallery-create-success' => '記事に動画ギャラリーを追加しました',
-	'vet-gallery-add-success' => '記事のギャラリーに動画を追加しました',
-	'vet-left' => '左寄せ',
-	'vet-right' => '右寄せ',
-	'vet-gallery' => 'ギャラリー',
-	'vet-protected' => 'この動画は保護されています',
-);
-
-/** Kalaallisut (kalaallisut)
- * @author Qaqqalik
- */
-$messages['kl'] = array(
-	'vet-find-btn' => 'Ujarlerit',
-);
-
-/** Kannada (ಕನ್ನಡ)
- * @author Nayvik
- */
-$messages['kn'] = array(
-	'vet-thiswiki' => 'ಈ ವಿಕಿ',
-	'vet-flickr' => 'ಫ಼್ಲಿಕ್ರ್',
-	'vet-upload-btn' => 'ಹೋಗು',
-	'vet-find-btn' => 'ಹುಡುಕು',
-	'vet-name' => 'ಹೆಸರು',
-	'vet-left' => 'ಎಡ',
-	'vet-right' => 'ಬಲ',
-);
-
-/** Colognian (Ripoarisch)
- * @author Purodha
- */
-$messages['ksh'] = array(
-	'vet-upload-btn' => 'Lohß Jonn!',
-);
-
-/** Kurdish (Latin script) (‪Kurdî (latînî)‬)
- * @author George Animal
- */
-$messages['ku-latn'] = array(
-	'vet-upload-btn' => ' Biçe',
-	'vet-find-btn' => 'Lêbigere',
-	'vet-name' => 'Nav',
-	'vet-show-message' => 'mesajê nîşan bide',
-	'vet-hide-message' => 'mesajê veşêre',
-	'vet-hide-license-msg' => 'Lîsansa nivîsê veşêre',
-	'vet-show-license-msg' => 'Lîsansa nivîsê nîşan bide',
-	'vet-left' => 'Çep',
-	'vet-right' => 'Rast',
-);
-
-/** Latin (Latina)
- * @author Rsa23899
- */
-$messages['la'] = array(
-	'vet-name' => 'Nōmen',
-	'vet-left' => 'Sinister',
-	'vet-right' => 'Dexter',
-);
-
-/** Luxembourgish (Lëtzebuergesch)
- * @author Robby
- */
-$messages['lb'] = array(
-	'vet-upload' => 'Komplett URL',
-	'vet-find' => 'E Video fannen',
-	'vet-thiswiki' => 'Dës Wiki',
-	'vet-upload-btn' => 'Weider',
-	'vet-find-btn' => 'Sichen',
-	'vet-rename' => 'Video ëmbenennen',
-	'vet-name' => 'Numm',
-	'vet-license' => 'Lizenz',
-	'vet-size' => 'Gréisst',
-	'vet-thumbnail' => 'Miniaturbild',
-	'vet-prev' => 'Vireg 8',
-	'vet-next' => 'Nächst 8',
-	'vet-back' => 'Zréck',
-	'vet-close' => 'Zoumaachen',
-	'vet-show-message' => 'Message weisen',
-	'vet-hide-message' => 'Message verstoppen',
-	'vet-add-button' => 'Video derbäisetzen',
-	'vet-left' => 'Lenks',
-	'vet-right' => 'Riets',
-	'vet-gallery' => 'Galerie',
-);
-
-/** Limburgish (Limburgs)
- * @author Ooswesthoesbes
- */
-$messages['li'] = array(
-	'vet-upload' => 'Gansen URL',
-);
-
-/** Macedonian (македонски)
- * @author Bjankuloski06
- */
-$messages['mk'] = array(
-	'vet-upload' => 'Полна URL-адреса',
-	'vet-notlogged' => 'Најавете се или регистрирајте сметка за да можете да вметнувате видеоснимки',
-	'vet-find' => 'Пронајди видеоснимка',
-	'vet-thiswiki' => 'Ова вики',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Следно',
-	'vet-find-btn' => 'Пребарај',
-	'vet-flickr-inf' => 'Пребарувајте правилно лиценцирани слободни фотографии од Flickr и увезувајте ги во вашето вики.',
-	'vet-conflict-inf' => '<h1>Упс, веќе постои видеоснимка со тоа име.</h1>Што би сакале да правите со <b>$1</b>?',
-	'vet-rename' => 'Преименувај видеоснимка',
-	'vet-existing' => 'Користи постоечка видеоснимка',
-	'vet-overwrite' => 'Замени ја постоечката видеоснимка со мојата',
-	'vet-insert' => 'Вметни',
-	'vet-details-inf' => '<h1>Податоци за податотеката</h1>Назначете име за оваа видеоснимка. Треба да биде описно.',
-	'vet-name' => 'Име',
-	'vet-license' => 'Лиценца',
-	'vet-details-inf2' => 'Нагодувања на приказот на видеоснимки',
-	'vet-size' => 'Големина',
-	'vet-thumbnail' => 'Минијатура',
-	'vet-fullsize' => 'Во полна големина ($1x$2px)',
-	'vet-width' => 'Ширина',
-	'vet-layout' => 'Распоред',
-	'vet-caption' => 'Опис',
-	'vet-optional' => '  (Незадолжително)',
-	'vet-insert2' => 'Вметни видеоснимка',
-	'vet-update' => 'Поднови видеоснимка',
-	'vet-video-replace-link' => 'За да ја замените оваа видеоснимка, одете на [[:$1]]',
-	'vet-prev' => 'Претходни 8',
-	'vet-next' => 'Следни 8',
-	'vet-metacafe' => 'Видеоснимки од Metacafe ($1 {{PLURAL:$1|резултат|резултати}})',
-	'vet-insert3' => 'Вметни ја видеоснимкава',
-	'vet-recent-inf' => 'Неодамна вметнати видеоснимки',
-	'vet-thiswiki2' => '!Видеоснимки на ова вики ($1 {{PLURAL:$1|резултат|резултати}})',
-	'vet-success' => '<h1>Усшешно извршено</h1>',
-	'vet-return' => 'Назад на уредувањето',
-	'vet-back' => 'назад',
-	'vet-close' => 'затвори',
-	'vet-warn1' => 'Најпрвин треба да назначите барање!',
-	'vet-warn2' => 'Наведете ја URL-адресата на видеоснимката што сакате да ја додадете',
-	'vet-warn3' => 'Наведете го името на видеоснимката што сакате да ја додадете',
-	'vet-imagelink' => '<b>Ново!</b> Вметнување на видеоснимки',
-	'vet-imagebutton' => 'Вметни видеоснимки',
-	'vet-license-cc' => 'Лиценцирано со Creative Commons Наведи извор-Сподели под исти услови 3.0',
-	'vet-name-incorrect' => 'Името на оваа видеоснимка содржи неважечки знаци, како на пр. #',
-	'vet-description' => 'Внесете ја полната URL-адреса од едно од поддржаните мрежни места. Вид.',
-	'vet-uploadtext' => "'''Овој образец служи за брзо подигање и поставување на податотеки во статии.'''
-
-''Погледајте го и [[Special:Log/upload|дневникот на подигања]] и [[Special:ImageList|списокот на сите подигнати податотеки]].''",
-	'vet-show-message' => 'прикажи порака',
-	'vet-hide-message' => 'скриј порака',
-	'vet-hide-license-msg' => 'скриј текст на лиценцата',
-	'vet-show-license-msg' => 'прикажи текст на лиценцата',
-	'vet-max-thumb' => 'пречекорена е максиналната дозволена големина на минијатурата; ја враќам првичната големина',
-	'vet-notallowed' => 'Не можете да вметнувате видеоснимки на ова вики',
-	'vet-title' => 'Додајте видеоснимка во вашата статија',
-	'vet-supported' => 'Поддржани мрежни места за видео:',
-	'vet-preview' => '<i>Овој преглед не е во вистински размер</i>',
-	'vet-bad-url' => 'Наведената URL-адреса е неважечка',
-	'vet-non-existing' => 'Наведената видеоснимка не постои',
-	'vet-main-info' => '<i>Внесете ја полната URL-адреса од едно од поддржаните мрежни места</i>',
-	'vet-page-success' => 'Видеостраницата е успешно подновена. Може да треба да ја исчистите видеостраницата за да можете да ги видите промените.',
-	'vet-add-button' => 'Додај видео',
-	'vet-added-from-gallery' => 'видеоснимката е додадена од видеогалеријата',
-	'vet-see-all' => 'сè',
-	'vet-bad-search' => 'Грешка при вчитувањето на бараното.',
-	'vet-no-preview' => 'Не можете да додавате видеоснимки од статија додека сте во прегледен режим',
-	'vet-single-success' => 'Видеоснимката е успешно поставена во оваа статија',
-	'vet-gallery-create-success' => 'Во оваа статија е додадена видеогалерија',
-	'vet-gallery-add-success' => 'Видеоснимката е додадена во видеогалеријата во оваа статија',
-	'vet-left' => 'Лево',
-	'vet-right' => 'Десно',
-	'vet-gallery' => 'Галерија',
-	'vet-protected' => 'Оваа видеоснимка е заштитена',
-);
-
-/** Mongolian (монгол)
- * @author Chinneeb
- */
-$messages['mn'] = array(
-	'vet-upload-btn' => 'Явах',
-);
-
-/** Malay (Bahasa Melayu)
- * @author Anakmalaysia
- */
-$messages['ms'] = array(
-	'vet-upload' => 'URL penuh',
-	'vet-notlogged' => 'Log masuk atau berdaftar untuk menyisipkan video',
-	'vet-find' => 'Cari video',
-	'vet-thiswiki' => 'Wiki ini',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Seterusnya',
-	'vet-find-btn' => 'Cari',
-	'vet-flickr-inf' => 'Cari gambar-gambar bebas dengan lesen yang sesuai dari Flickr, kemudian import gambar-gambar itu ke dalam wiki anda.',
-	'vet-conflict-inf' => '<h1>Alamak, nama itu sudah dipakai oleh video lain.</h1>Anda nak buat apa dengan <b>$1</b>?',
-	'vet-rename' => 'Tukar nama video',
-	'vet-existing' => 'Gunakan video sedia ada',
-	'vet-overwrite' => 'Gantikan video sedia ada dengan video anda',
-	'vet-insert' => 'Sisipkan',
-	'vet-details-inf' => '<h1>Butiran fail</h1>Berikan nama yang deskriptif kepada video ini.',
-	'vet-name' => 'Nama',
-	'vet-license' => 'Lesen',
-	'vet-details-inf2' => 'Pilihan paparan video',
-	'vet-size' => 'Saiz',
-	'vet-thumbnail' => 'Gambar kenit',
-	'vet-fullsize' => 'Saiz penuh ( $1 x $2 px)',
-	'vet-width' => 'Lebar',
-	'vet-layout' => 'Tataletak',
-	'vet-caption' => 'Kapsyen',
-	'vet-optional' => ' (Tidak wajib)',
-	'vet-insert2' => 'Sisipkan video',
-	'vet-update' => 'Kemaskini video',
-	'vet-video-replace-link' => 'Untuk mengganti video ini, sila ke [[:$1]]',
-	'vet-prev' => '8 yang sebelumnya',
-	'vet-next' => '8 yang berikutnya',
-	'vet-metacafe' => 'Video Metacafe ($1 hasil)',
-	'vet-insert3' => 'Sisipkan video ini',
-	'vet-recent-inf' => 'Video yang baru disisipkan',
-	'vet-thiswiki2' => 'Video wiki ini ($1 hasil)',
-	'vet-success' => '<h1>Berjaya</h1>',
-	'vet-return' => 'Kembali menyunting',
-	'vet-back' => 'Kembali',
-	'vet-close' => 'Tutup',
-	'vet-warn1' => 'Anda mesti menyatakan pertanyaan terlebih dahulu!',
-	'vet-warn2' => 'Sila nyatakan URL video yang hendak dibubuh',
-	'vet-warn3' => 'Sila nyatakan nama video yang hendak dibubuh',
-	'vet-imagelink' => '<b>Baru!</b> Sisipkan video',
-	'vet-imagebutton' => 'Sisipkan video',
-	'vet-license-cc' => 'Dilesenkan di bawah Creative Commons Attribution-Share Alike 3.0',
-	'vet-name-incorrect' => 'Nama video ini mengandungi aksara yang tidak boleh diterima, seperti #',
-	'vet-description' => 'Isikan URL dari mana-mana tapak yang disokong. Lihat',
-	'vet-uploadtext' => "'''Borang ini membolehkan anda memuat naik dan memaparkan fail pada rencana dengan segera.'''
-
-''Lihat juga [[Special:Log/upload|log muat naik]] dan [[Special:ImageList|senarai semua fail yang dimuat naik]].''",
-	'vet-show-message' => 'paparkan pesanan',
-	'vet-hide-message' => 'sorokkan pesanan',
-	'vet-hide-license-msg' => 'sorokkan teks lesen',
-	'vet-show-license-msg' => 'tunjukkan teks lesen',
-	'vet-max-thumb' => 'melebihi had saiz gambar kenit maksimum, dipulihkan ke saiz asal',
-	'vet-notallowed' => 'Anda tidak boleh menyisipkan video di wiki ini.',
-	'vet-title' => 'Sisipkan video pada rencana anda',
-	'vet-supported' => 'Laman video yang disokong:',
-	'vet-preview' => '<i>Pralihat ini tidak mengikut skala</i>',
-	'vet-bad-url' => 'URL yang diberikan tidak sah',
-	'vet-non-existing' => 'Video yang diberikan tidak wujud',
-	'vet-main-info' => '<i>Taipkan URL penuh dari mana-mana tapak yang disokong</i>',
-	'vet-page-success' => 'Laman video berjaya dikemaskini. Anda mungkin perlu membersihkan <i>cache</i> laman video untuk melihat perubahannya.',
-	'vet-add-button' => 'Sisipkan video',
-	'vet-added-from-gallery' => 'video disisipkan dari galeri video',
-	'vet-see-all' => 'semua',
-	'vet-bad-search' => 'Ralat ketika memuatkan pertanyaan carian.',
-	'vet-no-preview' => 'Anda tidak boleh menambahkan video daripada rencana dalam mod pralihat.',
-	'vet-single-success' => 'Video berjaya diletakkan dalam rencana ini.',
-	'vet-gallery-create-success' => 'Galeri video disisipkan pada rencana ini.',
-	'vet-gallery-add-success' => 'Video telah dimasukkan ke dalam galeri video dalam rencana ini.',
-	'vet-left' => 'Kiri',
-	'vet-right' => 'Kanan',
-	'vet-gallery' => 'Galeri',
-	'vet-protected' => 'Video ini dilindungi.',
-);
-
-/** Maltese (Malti)
- * @author Chrisportelli
- */
-$messages['mt'] = array(
-	'vet-upload-btn' => 'MUR',
-	'vet-close' => 'Agħlaq',
-);
-
-/** Burmese (မြန်မာဘာသာ)
- * @author Erikoo
- */
-$messages['my'] = array(
-	'vet-upload-btn' => 'သွား​ပါ​',
-	'vet-find-btn' => 'ရှာ​ဖွေ​ရန်​',
-	'vet-name' => 'အမည်',
-	'vet-license' => 'လိုင်စင်',
-	'vet-size' => 'အရွယ်အစား',
-	'vet-thumbnail' => 'နမူနာပုံငယ်',
-	'vet-width' => 'အကျယ်',
-	'vet-prev' => 'နောက် 8',
-	'vet-back' => 'နောက်',
-	'vet-close' => 'ပိတ်ရန်',
-	'vet-see-all' => 'ဆက်လက်ကြည့်ရန်',
-);
-
-/** Norwegian Bokmål (‪norsk (bokmål)‬)
- * @author Audun
- * @author Nghtwlkr
- */
-$messages['nb'] = array(
-	'vet-upload' => 'Full URL',
-	'vet-notlogged' => 'Logg inn eller registrer deg for å sette inn videoer',
-	'vet-find' => 'Finn en video',
-	'vet-thiswiki' => 'Denne Wikien',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Neste',
-	'vet-find-btn' => 'Søk',
-	'vet-flickr-inf' => 'Søk gjennom forsvarlig lisensierte gratisbilder fra Flickr og importer dem til wikien din.',
-	'vet-conflict-inf' => '<h1>Oops, en video med det navnet finnes allerede.</h1>Hva vil du gjøre med <b>$1</b>?',
-	'vet-rename' => 'Gi video nytt navn',
-	'vet-existing' => 'Bruk eksisterende video',
-	'vet-overwrite' => 'Overskriv den eksisterende videoen med din',
-	'vet-insert' => 'Sett inn',
-	'vet-details-inf' => '<h1>Fildetaljer</h1>Angi et navn for denne videoen. Gi den noe beskrivende.',
-	'vet-name' => 'Navn',
-	'vet-license' => 'Lisens',
-	'vet-details-inf2' => 'Videovisningsvalg',
-	'vet-size' => 'Størrelse',
-	'vet-thumbnail' => 'Miniatyrbilde',
-	'vet-fullsize' => 'Full størrelse ($1x$2px)',
-	'vet-width' => 'Bredde',
-	'vet-layout' => 'Oppsett',
-	'vet-caption' => 'Bildetekst',
-	'vet-optional' => '  (Valgfri)',
-	'vet-insert2' => 'Sett inn video',
-	'vet-update' => 'Oppdater video',
-	'vet-video-replace-link' => 'For å bytte ut denne videoen, vennligst gå til [[:$1]]',
-	'vet-prev' => 'Forrige 8',
-	'vet-next' => 'Neste 8',
-	'vet-metacafe' => 'Metacafe-videoer ({{PLURAL:$1|Ett resultat|$1 resultat}})',
-	'vet-insert3' => 'Sett inn denne videoen',
-	'vet-recent-inf' => 'Nylig innsatte videoer',
-	'vet-thiswiki2' => 'Denne wikiens videoer ({{PLURAL:$1|Ett resultat|$1 resultat}})',
-	'vet-success' => '<h1>Suksess</h1>',
-	'vet-return' => 'Tilbake til redigering',
-	'vet-back' => 'tilbake',
-	'vet-close' => 'lukk',
-	'vet-warn1' => 'Du må spesifisere en spørring først.',
-	'vet-warn2' => 'Vennligst oppgi nettadressen til videoen til vil legge til',
-	'vet-warn3' => 'Vennligst oppgi navnet på videoen du ønsker å legge til',
-	'vet-imagelink' => '<b>Ny!</b> Sett inn videoer',
-	'vet-imagebutton' => 'Sett inn videoer',
-	'vet-license-cc' => 'Lisensiert under Creative Commons Navngivelse-Del på samme vilkår 3.0',
-	'vet-name-incorrect' => 'Dette videonavnet inneholder ugyldige tegn, som #',
-	'vet-description' => 'Skriv inn hele nettadressen fra et av de støttede nettstedene. Se',
-	'vet-uploadtext' => "'''Dette skjemaet gjør at du raskt kan laste opp og vise filer i artikler.'''
-
-''Se også [[Special:Log/upload|opplastingsloggen]] og en [[Special:ImageList|liste over alle opplastede filer]].''",
-	'vet-show-message' => 'vis melding',
-	'vet-hide-message' => 'skjul melding',
-	'vet-hide-license-msg' => 'skjul lisenstekst',
-	'vet-show-license-msg' => 'vis lisenstekst',
-	'vet-max-thumb' => 'maks størrelse for miniatyrbilde overskredet, går tilbake til opprinnelig størrelse',
-	'vet-notallowed' => 'Du kan ikke sette inn videoer på denne wikien',
-	'vet-title' => 'Legg til en video i artikkelen din',
-	'vet-supported' => 'Støttede videosteder:',
-	'vet-preview' => '<i>Denne forhåndsvisningen er ikke i riktig skala</i>',
-	'vet-bad-url' => 'Den oppgitte URL-en er ugyldig',
-	'vet-non-existing' => 'Den oppgitte videoen finnes ikke',
-	'vet-main-info' => '<i>Skriv inn hele nettadressen fra noen av de støttede nettstedene</i>',
-	'vet-page-success' => 'Videosiden ble oppdatert. Du må kanskje slette mellomlagerfiler for videosiden for å se endringen.',
-	'vet-add-button' => 'Legg til video',
-	'vet-added-from-gallery' => 'video lagt til fra videogalleri',
-	'vet-see-all' => 'alle',
-	'vet-bad-search' => 'Feil under lasting av søkespørring.',
-	'vet-no-preview' => 'Du kan ikke legge til video fra artikkel i forhåndsvisningsmodus',
-	'vet-single-success' => 'Video ble vellykket plassert i denne artikkelen',
-	'vet-gallery-create-success' => 'Et videogalleri ble lagt til denne artikkelen',
-	'vet-gallery-add-success' => 'Videoen har blitt lagt til videogalleriet til denne artikkelen',
-	'vet-left' => 'Venstre',
-	'vet-right' => 'Høyre',
-	'vet-gallery' => 'Galleri',
-	'vet-protected' => 'Denne videoen er beskyttet',
-);
-
-/** Nepali (नेपाली)
- * @author RajeshPandey
- */
-$messages['ne'] = array(
-	'vet-find-btn' => 'खोज्नुहोस',
-);
-
-/** Dutch (Nederlands)
- * @author Naudefj
- * @author Siebrand
- */
-$messages['nl'] = array(
-	'vet-upload' => 'Volledige URL',
-	'vet-notlogged' => "Meld u aan of registreet een gebruiker om video's toe te voegen",
-	'vet-find' => 'Video zoeken',
-	'vet-thiswiki' => 'Deze wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Volgende',
-	'vet-find-btn' => 'Zoeken',
-	'vet-flickr-inf' => 'Zoek correct gelicenseerde vrije afbeeldingen van Flickr en importeer ze in uw wiki.',
-	'vet-conflict-inf' => '<h1>De video met deze naam bestaat al</h1>Wat wilt u met <b>$1</b> doen?',
-	'vet-rename' => 'Video hernoemen',
-	'vet-existing' => 'Bestaande video gebruiken',
-	'vet-overwrite' => 'De bestaande video met uw bestand overschrijven',
-	'vet-insert' => 'Invoegen',
-	'vet-details-inf' => '<h1>Bestandsgegevens</h1>
-Geef een beschrijvende naam op voor deze video.',
-	'vet-name' => 'Naam',
-	'vet-license' => 'Licentie',
-	'vet-details-inf2' => 'Videoweergaveinstellingen',
-	'vet-size' => 'Grootte',
-	'vet-thumbnail' => 'Miniatuur',
-	'vet-fullsize' => 'Volledige grootte ($1x$2px)',
-	'vet-width' => 'Breedte',
-	'vet-layout' => 'Vormgeving',
-	'vet-caption' => 'Bijschrift',
-	'vet-optional' => ' (optioneel)',
-	'vet-insert2' => 'Video invoegen',
-	'vet-update' => 'Video bijwerken',
-	'vet-video-replace-link' => 'Ga alstublieft naar [[:$1]] om deze video te vervangen',
-	'vet-prev' => 'Vorige 8',
-	'vet-next' => 'Volgende 8',
-	'vet-metacafe' => "Metcafe-video's ($1 {{PLURAL:$1|resultaat|resultaten}})",
-	'vet-insert3' => 'Deze video invoegen',
-	'vet-recent-inf' => "Recent toegevoegde video's",
-	'vet-thiswiki2' => "Video's van deze wiki ($1 {{PLURAL:$1|resultaat|resultaten}})",
-	'vet-success' => '<h1>Geslaagd</h1>',
-	'vet-return' => 'Terug naar bewerken',
-	'vet-back' => 'terug',
-	'vet-close' => 'sluiten',
-	'vet-warn1' => 'U moet eerst een zoekopdracht ingeven!',
-	'vet-warn2' => 'Geef alstublieft de URL op van de video die u wilt toevoegen',
-	'vet-warn3' => 'Geef alstublieft de naam op van de video die u wilt toevoegen',
-	'vet-imagelink' => '<b>Nieuw!</b> Video toevoegen',
-	'vet-imagebutton' => "Video's toevoegen",
-	'vet-license-cc' => 'Gelicenseerd onder de Creative Commons-licentie Naamsvermelding-Gelijk delen 3.0',
-	'vet-name-incorrect' => 'De naam van deze video bevat ongeldige tekens, zoals #.',
-	'vet-description' => 'Geef een volledige URL op van een van de ondersteunde websites. Zie',
-	'vet-uploadtext' => "'''Via dit formulier kunt u snel bestanden uploaden en in pagina's plaatsen.'''
-
-''Zie ook het [[Special:Log/upload|uploadlogboek]] en een [[Special:ImageList|lijst met alle toegevoegde bestanden]].''",
-	'vet-show-message' => 'bericht weergeven',
-	'vet-hide-message' => 'bericht verbergen',
-	'vet-hide-license-msg' => 'licentietekst verbergen',
-	'vet-show-license-msg' => 'licentietekst weergeven',
-	'vet-max-thumb' => 'de maximale miniatuurgrootte is overschreden; de originele afmetingen worden ingesteld',
-	'vet-notallowed' => "U kunt geen video's toevoegen in deze wiki",
-	'vet-title' => 'Video aan uw pagina toevoegen',
-	'vet-supported' => 'Ondersteunde videosites:',
-	'vet-preview' => '<i>Deze voorvertoning is niet correct geschaald</i>',
-	'vet-bad-url' => 'De opgegeven URL is ongeldig',
-	'vet-non-existing' => 'De opgegeven video bestaat niet',
-	'vet-main-info' => '<i>Geef een volledige URL op van een van de ondersteunde websites.</i>',
-	'vet-page-success' => 'Die videopagina is bijgewerkt.
-Wellicht moet u de videopagina opschonen om de wijzigingen te zien.',
-	'vet-add-button' => 'Video toevoegen',
-	'vet-added-from-gallery' => 'video toegevoegd vanuit videogalerij',
-	'vet-see-all' => 'alle',
-	'vet-bad-search' => 'Er is een fout opgetreden bij het laden van de zoekopdracht.',
-	'vet-no-preview' => 'U kunt geen video uit een pagina toevoegen in de voorvertoning',
-	'vet-single-success' => 'De video is aan de pagina toegevoegd.',
-	'vet-gallery-create-success' => 'De videogalerij is aan deze pagina toegevoegd',
-	'vet-gallery-add-success' => 'De video is toegevoegd aan de galerij op deze pagina.',
-	'vet-left' => 'Links',
-	'vet-right' => 'Rechts',
-	'vet-gallery' => 'Galerij',
-	'vet-protected' => 'Deze video is beveiligd',
-);
-
-/** ‪Nederlands (informeel)‬ (‪Nederlands (informeel)‬)
- * @author Siebrand
- */
-$messages['nl-informal'] = array(
-	'vet-flickr-inf' => 'Zoek correct gelicenseerde vrije afbeeldingen van Flickr en importeer ze in je wiki.',
-	'vet-conflict-inf' => '<h1>De video met deze naam bestaat al</h1>Wat wil je met <b>$1</b> doen?',
-	'vet-overwrite' => 'De bestaande video met jouw bestand overschrijven',
-	'vet-warn1' => 'Je moet eerst een zoekopdracht ingeven!',
-	'vet-warn2' => 'Geef alstublieft de URL op van de video die je wilt toevoegen',
-	'vet-warn3' => 'Geef alsjeblieft de naam op van de video die je wilt toevoegen',
-	'vet-uploadtext' => "'''Via dit formulier kan je snel bestanden uploaden en in pagina's plaatsen.'''
-
-''Zie ook het [[Special:Log/upload|uploadlogboek]] en een [[Special:ImageList|lijst met alle toegevoegde bestanden]].''",
-	'vet-notallowed' => "Je kunt geen video's toevoegen in deze wiki",
-	'vet-title' => 'Video aan je pagina toevoegen',
-	'vet-page-success' => 'Die videopagina is bijgewerkt.
-Wellicht moet je de videopagina opschonen om de wijzigingen te zien.',
-	'vet-no-preview' => 'Je kunt geen video uit een pagina toevoegen in de voorvertoning',
-);
-
-/** Occitan (occitan)
- * @author Cedric31
- */
-$messages['oc'] = array(
-	'vet-upload' => 'URL completa',
-	'vet-find' => 'Cercar una vidèo',
-	'vet-thiswiki' => 'Aqueste wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Apondre',
-	'vet-find-btn' => 'Cercar',
-	'vet-conflict-inf' => '<h1>Ops, existís ja una vidèo amb aqueste nom.</h1>Qué volètz far amb <b>$1</b>?',
-	'vet-rename' => 'Tornar nomenar la vidèo',
-	'vet-insert' => 'Inserir',
-	'vet-name' => 'Nom',
-	'vet-license' => 'Licéncia',
-	'vet-details-inf2' => "Las opcions d'afichatge vidèo",
-	'vet-size' => 'Talha',
-	'vet-thumbnail' => 'Apercebut',
-	'vet-width' => 'largor',
-	'vet-layout' => 'Orientacion',
-	'vet-caption' => 'Descripcion',
-	'vet-optional' => '(Opcional)',
-);
-
-/** Deitsch (Deitsch)
- * @author Xqt
- */
-$messages['pdc'] = array(
-	'vet-find-btn' => 'Uffgucke',
-	'vet-next' => 'Neegschte 8',
-	'vet-back' => 'zerrick',
-	'vet-close' => 'zumache',
-);
-
-/** Pälzisch (Pälzisch)
- * @author Manuae
- */
-$messages['pfl'] = array(
-	'vet-upload-btn' => 'Negschde',
-	'vet-find-btn' => 'Suche',
-	'vet-name' => 'Noame',
-	'vet-size' => 'Greeß',
-	'vet-prev' => 'Voriche 8',
-	'vet-next' => 'Negschte 8',
-	'vet-back' => 'zrigg',
-	'vet-see-all' => 'alle',
-	'vet-left' => 'Lings',
-	'vet-right' => 'Reschds',
-	'vet-gallery' => 'Galarie',
-);
-
-/** Polish (polski)
- * @author BeginaFelicysym
- * @author Sp5uhe
- */
-$messages['pl'] = array(
-	'vet-upload' => 'Pełny adres URL',
-	'vet-notlogged' => 'Zaloguj się lub zarejestruj aby dodać film',
-	'vet-find' => 'Znajdź film',
-	'vet-thiswiki' => 'Ta wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Dalej',
-	'vet-find-btn' => 'Szukaj',
-	'vet-flickr-inf' => 'Znajdź w serwisie Flickr pliki na odpowiedniej licencji i umieść je na wiki.',
-	'vet-conflict-inf' => '<h1>Ups, film o tej nazwie już istnieje.</h1>Co chcesz zrobić z <b>$1</b>?',
-	'vet-rename' => 'Zmień nazwę filmu',
-	'vet-existing' => 'Użyj istniejący film',
-	'vet-overwrite' => 'Nadpisz istniejący film',
-	'vet-insert' => 'Wstaw',
-	'vet-details-inf' => '<h1>Szczegóły pliku</h1>
-Podaj nazwę tego filmu. Może to być np. opis jego zawartości.',
-	'vet-name' => 'Nazwa',
-	'vet-license' => 'Licencja',
-	'vet-details-inf2' => 'Opcje wyświetlania filmu',
-	'vet-size' => 'Rozmiar',
-	'vet-thumbnail' => 'Miniatura',
-	'vet-fullsize' => 'Pełnowymiarowy ($1x$2pks.)',
-	'vet-width' => 'Szerokość',
-	'vet-layout' => 'Układ',
-	'vet-caption' => 'Podpis',
-	'vet-optional' => '(Opcjonalnie)',
-	'vet-insert2' => 'Wstaw film',
-	'vet-update' => 'Aktualizuj film',
-	'vet-video-replace-link' => 'Jeśli chcesz zastąpić ten film, przejdź do [[:$1]]',
-	'vet-prev' => 'Poprzednie 8',
-	'vet-next' => 'Następne 8',
-	'vet-metacafe' => 'Filmy z Metacafe ($1 {{PLURAL:$1|rezultat|rezultaty|rezultatów}})',
-	'vet-insert3' => 'Wstaw ten film',
-	'vet-recent-inf' => 'Niedawno wstawione filmy',
-	'vet-thiswiki2' => 'Filmy w tej wiki ($1 {{PLURAL:$1|rezultat|rezultaty|rezultatów}})',
-	'vet-success' => '<h1>Sukces</h1>',
-	'vet-return' => 'Powrót do edycji',
-	'vet-back' => 'Wstecz',
-	'vet-close' => 'Zamknij',
-	'vet-warn1' => 'Należy najpierw określić zapytanie!',
-	'vet-warn2' => 'Podaj adres URL filmu, który chcesz dodać',
-	'vet-warn3' => 'Podaj nazwę filmu, który chcesz dodać',
-	'vet-imagelink' => '<b>Nowe!</b> Wstawianie filmów',
-	'vet-imagebutton' => 'Wstaw filmy',
-	'vet-license-cc' => 'Udostępniane na licencji Creative Commons Uznanie autorstwa – Na tych samych warunkach 3.0',
-	'vet-name-incorrect' => 'Nazwa filmu zawiera niedozwolone znaki, jak #',
-	'vet-description' => 'Wprowadź pełny adres URL jednej z obsługiwanych witryn. Zobacz',
-	'vet-uploadtext' => "'''Ten formularz umożliwia szybkie przesyłanie i zamieszczanie plików w artykułach.'''
-
-''Zobacz również [[Special:Log/upload|rejestr przesyłania]] oraz [[Special:ImageList|spis wszystkich przesłanych plików]].''",
-	'vet-show-message' => 'pokaż wiadomość',
-	'vet-hide-message' => 'ukryj wiadomość',
-	'vet-hide-license-msg' => 'ukryj treść licencji',
-	'vet-show-license-msg' => 'pokaż treść licencji',
-	'vet-max-thumb' => 'przekroczony maksymalny rozmiar dla miniatury, przywrócono wielkość pierwotną',
-	'vet-notallowed' => 'Nie masz uprawnień do dodawania filmów na tej wiki.',
-	'vet-title' => 'Dodaj film do swojego artykułu',
-	'vet-supported' => 'Wspierane strony z filmami:',
-	'vet-preview' => '<i>Podgląd nie został wyskalowany</i>',
-	'vet-bad-url' => 'Podany adres URL jest nieprawidłowy',
-	'vet-non-existing' => 'Podany film nie istnieje',
-	'vet-main-info' => '<i>Wpisz pełny adres URL jednej z obsługiwanych witryn</i>',
-	'vet-page-success' => 'Strona filmu została zaktualizowana. Być może trzeba będzie odświeżyć stronę, aby zobaczyć zmianę.',
-	'vet-add-button' => 'Dodaj film',
-	'vet-added-from-gallery' => 'dodano film z galerii',
-	'vet-see-all' => 'wszystko',
-	'vet-bad-search' => 'Błąd podczas ładowania zapytania.',
-	'vet-no-preview' => 'Nie możesz w trybie podglądu dodać filmu z artykułu.',
-	'vet-single-success' => 'Film został pomyślnie dodany do artykułu.',
-	'vet-gallery-create-success' => 'Galeria filmów została dodana do artykułu.',
-	'vet-gallery-add-success' => 'Film został dodany do galerii w artykule',
-	'vet-left' => 'Lewo',
-	'vet-right' => 'Prawo',
-	'vet-gallery' => 'Galeria',
-	'vet-protected' => 'Ten film jest zabezpieczony.',
-);
-
-/** Piedmontese (Piemontèis)
- * @author Borichèt
- * @author Dragonòt
- */
-$messages['pms'] = array(
-	'vet-upload' => 'Adrëssa dla ragnà completa',
-	'vet-notlogged' => "Ch'a intra ant ël sistema o ch'as argistra për anserì dij filmà",
-	'vet-find' => 'Trové un filmà',
-	'vet-thiswiki' => 'Sta Wiki-sì',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Apress',
-	'vet-find-btn' => 'Sërché',
-	'vet-flickr-inf' => 'Sërché che fòto lìbere licensià për da bin su Flickr e amporteje su soa wiki.',
-	'vet-conflict-inf' => "<h1>Contacc, un filmà con col nòm a esist già.</h1>Lòn ch'a veul fé con <b>$1</b>?",
-	'vet-rename' => 'Arnominé ël filmà',
-	'vet-existing' => 'Dovré un filmà Esistent',
-	'vet-overwrite' => 'Coaté ël filmà esistent con ël sò',
-	'vet-insert' => 'Anserì',
-	'vet-details-inf' => "<h1>Detaj dl'archivi</h1>Ch'a spessìfica un nòm për ës filmà-sì. Ch'a sia cheicòs ëd descritiv.",
-	'vet-name' => 'Nòm',
-	'vet-license' => 'Licensa',
-	'vet-details-inf2' => 'Opsion ëd visualisassion dël filmà',
-	'vet-size' => 'Dimension',
-	'vet-thumbnail' => 'Figurin-a',
-	'vet-fullsize' => 'Dimension pien-a ($1x$2pt)',
-	'vet-width' => 'Larghëssa',
-	'vet-layout' => 'Presentassion',
-	'vet-caption' => 'Descrission',
-	'vet-optional' => ' (Opsional)',
-	'vet-insert2' => 'Anserì ël filmà',
-	'vet-update' => 'Modifiché ël filmà',
-	'vet-video-replace-link' => "Për rimpiassé ës filmà-sì, për piasì ch'a vada a [[:$1]]",
-	'vet-prev' => '8 prima',
-	'vet-next' => '8 dapress',
-	'vet-metacafe' => 'Filmà ëd Metacafè ($1 {{PLURAL:$1|arzultà|arzultà}})',
-	'vet-insert3' => 'Anserì ës filmà',
-	'vet-recent-inf' => 'Filmà giontà ëd recent',
-	'vet-thiswiki2' => 'Filmà dë sta Wiki-sì ($1 {{PLURAL:$1|arzultà|arzultà}})',
-	'vet-success' => '<h1>Fàit</h1>',
-	'vet-return' => 'Artorné a le modìfiche',
-	'vet-back' => 'andré',
-	'vet-close' => 'sara',
-	'vet-warn1' => "A deuv prima spessifiché l'arcesta!",
-	'vet-warn2' => "Për piasì, ch'a buta l'adrëssa an sl'aragnà dël filmà ch'a veul gionté",
-	'vet-warn3' => "Për piasì, ch'a buta ël nòm dël filmà ch'a veul gionté",
-	'vet-imagelink' => '<b>Neuv!</b> Anserì dij filmà',
-	'vet-imagebutton' => 'Gionté dij filmà',
-	'vet-license-cc' => 'Licensià sota Creative Commons Attribution-Share Alike 3.0',
-	'vet-name-incorrect' => 'Ël nòm ëd cost filmà a conten dij caràter pa bon, com #',
-	'vet-description' => "Ch'a anserissa l'anliura completa da qualsëssìa dij sit mantnù. Vëdde",
-	'vet-uploadtext' => "'''Cost formolari a-j përmët ëd carié e visualisé an pressa dj'archivi ant j'artìcoj.'''
-
-''Ch'a bèica ëdcò ël [[Special:Log/upload|registr dij cariagi]] e na [[Special:ImageList|lista ëd tùit j'archivi carià]].''",
-	'vet-show-message' => 'smon-e ël mëssagi',
-	'vet-hide-message' => 'stërma mëssagi',
-	'vet-hide-license-msg' => 'stërmé ël test ëd la licensa',
-	'vet-show-license-msg' => 'mostré ël test ëd la licensa',
-	'vet-max-thumb' => 'dimension màssima dla figurin-a sorpassà, artorn a la dimension original',
-	'vet-notallowed' => 'A peule pa anserì dij filmà su sta wiki-sì',
-	'vet-title' => 'Gionté un filmà a sò artìcol',
-	'vet-supported' => 'Sit ëd filmà sostnù:',
-	'vet-preview' => "<i>Costa preuva a l'é nen an scala</i>",
-	'vet-bad-url' => "L'adrëssa dl'aragnà butà a l'é pa bon-a",
-	'vet-non-existing' => 'Ël filmà giontà a esist pa',
-	'vet-main-info' => "<i>Ch'a anserissa l'anliura completa da qualsëssìa dij sit mantnù. </i>",
-	'vet-page-success' => "La pàgina dël filmà a l'é stàita modificà për da bin. A dovrìa purghé la pàgina dël filmà për vëdde ij cangiament.",
-	'vet-add-button' => 'Gionté un filmà',
-	'vet-added-from-gallery' => 'filmà giontà da na galarìa ëd filmà',
-	'vet-see-all' => 'tuti',
-	'vet-bad-search' => "Eror cariand l'arcesta d'arserca",
-	'vet-no-preview' => 'A peul pa gionté ëd filmà da artìcoj an manera preuva',
-	'vet-single-success' => "Ël filmà a l'é stàit piassà da bin an st'artìcol-sì",
-	'vet-gallery-create-success' => "Na galarìa ëd filmà a l'é stàita giontà a cost artìcol",
-	'vet-gallery-add-success' => "Un filmà a l'é stàit giontà a la galarìa ëd filmà an st'artìcol-sì",
-	'vet-left' => 'Snist',
-	'vet-right' => 'Drit',
-	'vet-gallery' => 'Galarìa',
-	'vet-protected' => "Ës filmà-sì a l'é protegiù",
-);
-
-/** Pashto (پښتو)
- * @author Ahmed-Najib-Biabani-Ibrahimkhel
- */
-$messages['ps'] = array(
-	'vet-upload' => 'بشپړ يو آر اېل',
-	'vet-find' => 'يوه ويډيو موندل',
-	'vet-thiswiki' => 'همدا ويکي',
-	'vet-flickr' => 'فلېکر',
-	'vet-upload-btn' => 'راتلونکی',
-	'vet-find-btn' => 'پلټل',
-	'vet-name' => 'نوم',
-	'vet-license' => 'منښتليک',
-	'vet-size' => 'کچه',
-	'vet-thumbnail' => 'بټنوک',
-	'vet-layout' => 'ليدنمخ',
-	'vet-close' => 'تړل',
-	'vet-show-message' => 'پيغام ښکاره کول',
-	'vet-hide-message' => 'پيغام پټول',
-	'vet-hide-license-msg' => 'د منښتليک متن پټول',
-	'vet-show-license-msg' => 'د منښتليک متن ښکاره کول',
-	'vet-title' => 'خپلې ليکنې کې يوه ويډيو ورګډول',
-	'vet-add-button' => 'ويډيو ورګډول',
-	'vet-see-all' => 'ټول',
-	'vet-left' => 'کيڼ',
-	'vet-right' => 'ښي',
-	'vet-gallery' => 'انځورتون',
-	'vet-protected' => 'دا ويډيو ژغورل شوې',
-);
-
-/** Portuguese (português)
- * @author Giro720
- * @author Hamilton Abreu
- */
-$messages['pt'] = array(
-	'vet-upload' => 'URL completa',
-	'vet-notlogged' => 'Para inserir vídeos, autentique-se ou registe-se',
-	'vet-find' => 'Procure um vídeo',
-	'vet-thiswiki' => 'Esta wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Seguinte',
-	'vet-find-btn' => 'Pesquisar',
-	'vet-flickr-inf' => 'Procure fotografias devidamente licenciadas do Flickr e importe-as para a sua wiki.',
-	'vet-conflict-inf' => '<h1>Já existe um vídeo com esse nome. </h1> O que é que quer fazer com o <b>$1</b>?',
-	'vet-rename' => 'Alterar o nome do vídeo',
-	'vet-existing' => 'Usar vídeo existente',
-	'vet-overwrite' => 'Sobrepor o seu vídeo ao existente',
-	'vet-insert' => 'Inserir',
-	'vet-details-inf' => '<h1>Detalhes do ficheiro</h1>Introduza um nome descritivo para este vídeo.',
-	'vet-name' => 'Nome',
-	'vet-license' => 'Licença',
-	'vet-details-inf2' => 'Opções de visionamento de vídeos',
-	'vet-size' => 'Dimensão',
-	'vet-thumbnail' => 'Miniatura',
-	'vet-fullsize' => 'Dimensão completa ($1x$2px)',
-	'vet-width' => 'Largura',
-	'vet-layout' => 'Posição',
-	'vet-caption' => 'Legenda',
-	'vet-optional' => '(Opcional)',
-	'vet-insert2' => 'Inserir vídeo',
-	'vet-update' => 'Actualizar vídeo',
-	'vet-video-replace-link' => 'Para substituir este vídeo, visite [[:$1]]',
-	'vet-prev' => '8 anteriores',
-	'vet-next' => '8 seguintes',
-	'vet-metacafe' => 'Vídeos Metacafe ($1 {{PLURAL:$1|resultado|resultados}})',
-	'vet-insert3' => 'Inserir este vídeo',
-	'vet-recent-inf' => 'Vídeos inseridos recentemente',
-	'vet-thiswiki2' => 'Vídeos desta wiki ($1 {{PLURAL:$1|resultado|resultados}})',
-	'vet-success' => '<h1>Sucesso</h1>',
-	'vet-return' => 'Voltar à edição',
-	'vet-back' => 'voltar',
-	'vet-close' => 'fechar',
-	'vet-warn1' => 'Primeiro tem de especificar uma consulta!',
-	'vet-warn2' => 'Introduza a URL do vídeo que pretende adicionar, por favor',
-	'vet-warn3' => 'Introduza o nome do vídeo que pretende adicionar, por favor',
-	'vet-imagelink' => '<b>Novo!</b> Inserir vídeos',
-	'vet-imagebutton' => 'Inserir vídeos',
-	'vet-license-cc' => 'Sujeito aos termos da licença Creative Commons - Atribuição - Partilha nos Mesmos Termos 3.0',
-	'vet-name-incorrect' => 'O nome deste vídeo contém caracteres inválidos',
-	'vet-description' => 'Introduza a URL completa de qualquer um dos sites suportados. Ver',
-	'vet-uploadtext' => "'''Este formulário permite-lhe carregar e incorporar ficheiros nas páginas de forma rápida.'''
-
-''Veja também o [[Special:Log/upload|registo de uploads]] e uma [[Special:ImageList|lista de todos os ficheiros carregados]].''",
-	'vet-show-message' => 'mostrar mensagem',
-	'vet-hide-message' => 'esconder mensagem',
-	'vet-hide-license-msg' => 'esconder texto da licença',
-	'vet-show-license-msg' => 'mostrar texto da licença',
-	'vet-max-thumb' => 'foi excedida a dimemsão máxima para miniaturas; a reverter para a original',
-	'vet-notallowed' => 'Não pode inserir vídeos nesta wiki',
-	'vet-title' => 'Adicione um vídeo à sua página',
-	'vet-supported' => 'Sites de vídeo suportados:',
-	'vet-preview' => '<i>Esta antevisão não tem uma escala proporcional</i>',
-	'vet-bad-url' => 'A URL fornecida é inválida',
-	'vet-non-existing' => 'O vídeo fornecido não existe',
-	'vet-main-info' => '<i>Introduza a URL completa de qualquer um dos sites suportados</i>',
-	'vet-page-success' => 'A página do vídeo foi actualizada. Pode ter de refrescar a página do vídeo para ver a alteração.',
-	'vet-add-button' => 'Adicionar vídeo',
-	'vet-added-from-gallery' => 'vídeo adicionado a partir da galeria de vídeo',
-	'vet-see-all' => 'todos',
-	'vet-bad-search' => 'Erro ao carregar a pesquisa.',
-	'vet-no-preview' => 'Não pode adicionar vídeo a partir de uma página em modo de antevisão',
-	'vet-single-success' => 'O vídeo foi colocado na página',
-	'vet-gallery-create-success' => 'Foi adicionada uma galeria de vídeo à página',
-	'vet-gallery-add-success' => 'O vídeo foi adicionado à galeria de vídeo da página',
-	'vet-left' => 'Esquerda',
-	'vet-right' => 'Direita',
-	'vet-gallery' => 'Galeria',
-	'vet-protected' => 'Este vídeo está protegido.',
-);
-
-/** Brazilian Portuguese (português do Brasil)
- * @author Giro720
- * @author Luckas Blade
- * @author Pedroca cerebral
- */
-$messages['pt-br'] = array(
-	'vet-upload' => 'URL completa',
-	'vet-notlogged' => 'Para inserir vídeos, autentique-se ou registre-se',
-	'vet-find' => 'Procure um vídeo',
-	'vet-thiswiki' => 'Esta wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Inserir',
-	'vet-find-btn' => 'Pesquisar',
-	'vet-flickr-inf' => 'Procure fotografias devidamente licenciadas do Flickr e importe-as para a sua wiki.',
-	'vet-conflict-inf' => '<h1>Já existe um vídeo com esse nome. </h1> O que você quer fazer com <b>$1</b>?',
-	'vet-rename' => 'Renomear vídeo',
-	'vet-existing' => 'Usar vídeo existente',
-	'vet-overwrite' => 'Substituir o vídeo já existente pelo seu vídeo',
-	'vet-insert' => 'Inserir',
-	'vet-details-inf' => '<h1>Detalhes do arquivo</h1>Introduza um nome descritivo para este vídeo.',
-	'vet-name' => 'Nome',
-	'vet-license' => 'Licença',
-	'vet-details-inf2' => 'Opções de visualização de vídeos',
-	'vet-size' => 'Tamanho',
-	'vet-thumbnail' => 'Miniatura',
-	'vet-fullsize' => 'Tamanho original ($1x$2px)',
-	'vet-width' => 'Largura',
-	'vet-layout' => 'Posição',
-	'vet-caption' => 'Legenda',
-	'vet-optional' => ' (Opcional)',
-	'vet-insert2' => 'Inserir vídeo',
-	'vet-update' => 'Atualizar vídeo',
-	'vet-video-replace-link' => 'Para substituir este vídeo, visite [[:$1]]',
-	'vet-prev' => '8 anteriores',
-	'vet-next' => '8 seguintes',
-	'vet-metacafe' => 'Vídeos Metacafe ($1 {{PLURAL:$1|resultado|resultados}})',
-	'vet-insert3' => 'Inserir este vídeo',
-	'vet-recent-inf' => 'Vídeos inseridos recentemente',
-	'vet-thiswiki2' => 'Vídeos desta wiki ($1 {{PLURAL:$1|resultado|resultados}})',
-	'vet-success' => '<h1>Sucesso</h1>',
-	'vet-return' => 'Voltar à edição',
-	'vet-back' => 'voltar',
-	'vet-close' => 'fechar',
-	'vet-warn1' => 'Você deve primeiro especificar uma consulta!',
-	'vet-warn2' => 'Introduza a URL do vídeo que você pretende adicionar, por favor',
-	'vet-warn3' => 'Introduza o nome do vídeo que você pretende adicionar, por favor',
-	'vet-imagelink' => '<b>Novo!</b> Inserir vídeos',
-	'vet-imagebutton' => 'Inserir vídeos',
-	'vet-license-cc' => 'Licenciado sob os termos da licença Creative Commons - Atribuição - Partilha nos Mesmos Termos 3.0',
-	'vet-name-incorrect' => 'O nome deste vídeo contém caracteres inválidos',
-	'vet-description' => 'Escreva a URL completa de qualquer um dos sites suportados. Ver',
-	'vet-uploadtext' => "'''Este formulário permite-lhe carregar e incorporar arquivos nas páginas de forma rápida.'''
-
-''Veja também o [[Special:Log/upload|registro de uploads]] e uma [[Special:ImageList|lista de todos os arquivos carregados]].''",
-	'vet-show-message' => 'mostrar mensagem',
-	'vet-hide-message' => 'ocultar mensagem',
-	'vet-hide-license-msg' => 'esconder texto da licença',
-	'vet-show-license-msg' => 'mostrar texto da licença',
-	'vet-max-thumb' => 'foi excedido o tamanho máximo para miniaturas; revertendo ao tamanho original',
-	'vet-notallowed' => 'Você não pode inserir vídeos nesta wiki',
-	'vet-title' => 'Adicionar um vídeo ao seu artigo',
-	'vet-supported' => 'Sites de vídeo suportados:',
-	'vet-preview' => '<i>Esta pré-visualização não tem uma escala proporcional</i>',
-	'vet-bad-url' => 'A URL fornecida é inválida',
-	'vet-non-existing' => 'O vídeo fornecido não existe',
-	'vet-main-info' => '<i>Procure um vídeo ou introduza uma URL completa de um dos sites suportados</i>',
-	'vet-page-success' => 'A página do vídeo foi atualizada. Você deve recarregar a página do vídeo para ver as alterações.',
-	'vet-add-button' => 'Adicionar vídeo',
-	'vet-added-from-gallery' => 'vídeo adicionado a partir da galeria de vídeo',
-	'vet-see-all' => 'ver todos',
-	'vet-bad-search' => 'Erro ao carregar a pesquisa.',
-	'vet-no-preview' => 'Você não pode adicionar vídeo a partir de uma página em modo de pré-visualização',
-	'vet-single-success' => 'O vídeo foi colocado na página',
-	'vet-gallery-create-success' => 'Uma galeria de vídeo foi adicionada a este artigo',
-	'vet-gallery-add-success' => 'O vídeo foi adicionado à galeria de vídeo desta página com sucesso.',
-	'vet-left' => 'Esquerda',
-	'vet-right' => 'Direita',
-	'vet-gallery' => 'Galeria',
-	'vet-protected' => 'Este vídeo está protegido.',
-);
-
-/** Romanian (română)
- * @author Stelistcristi
- */
-$messages['ro'] = array(
-	'vet-flickr' => 'Flickr',
-	'vet-name' => 'Nume',
-	'vet-license' => 'Licenţă',
-);
-
-/** Russian (русский)
- * @author Eleferen
- * @author Kaganer
- * @author Kuzura
- * @author Lockal
- * @author Александр Сигачёв
- */
-$messages['ru'] = array(
-	'vet-upload' => 'Полный URL',
-	'vet-notlogged' => 'Представьтесь системе или зарегистрируйтесь, чтобы вставлять видео',
-	'vet-find' => 'Найти видео',
-	'vet-thiswiki' => 'Эта вики',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Далее',
-	'vet-find-btn' => 'Найти',
-	'vet-flickr-inf' => 'Найдите свободные изображения с подходящей лицензией на Flickr и импортируйте их в свою вики.',
-	'vet-conflict-inf' => '<h1>К сожалению, видео с таким именем уже существует.</h1>Что вы хотите сделать с <b>$1</b>?',
-	'vet-rename' => 'Переименовать видео',
-	'vet-existing' => 'Использовать существующее видео',
-	'vet-overwrite' => 'Перезаписать существующее видео вашим',
-	'vet-insert' => 'Вставить',
-	'vet-details-inf' => '<h1>Имя файла</h1>Укажите имя для этого видео. Постарайтесь описать в нём содержимое файла.',
-	'vet-name' => 'Имя',
-	'vet-license' => 'Лицензия',
-	'vet-details-inf2' => 'Параметры отображения видео',
-	'vet-size' => 'Размер',
-	'vet-thumbnail' => 'Миниатюра',
-	'vet-fullsize' => 'Полный размер ($1x$2 пикс.)',
-	'vet-width' => 'Ширина',
-	'vet-layout' => 'Выравнивание',
-	'vet-caption' => 'Подпись',
-	'vet-optional' => ' (необязательно)',
-	'vet-insert2' => 'Вставить видео',
-	'vet-update' => 'Обновить видео',
-	'vet-video-replace-link' => 'Чтобы заменить это видео, перейдите на страницу [[:$1]]',
-	'vet-prev' => 'Предыдущие 8',
-	'vet-next' => 'Следующие 8',
-	'vet-metacafe' => 'Видео Metacafe ($1 {{PLURAL:$1|результат|результатов}})',
-	'vet-insert3' => 'Вставить это видео',
-	'vet-recent-inf' => 'Недавно добавленные видео',
-	'vet-thiswiki2' => 'Видео этой вики ($1 {{PLURAL:$1|результат|результатов}})',
-	'vet-success' => '<h1>Успешно</h1>',
-	'vet-return' => 'Возврат к редактированию',
-	'vet-back' => 'назад',
-	'vet-close' => 'закрыть',
-	'vet-warn1' => 'Вам нужно сначала указать запрос!',
-	'vet-warn2' => 'Приведите, пожалуйста, адрес видео, которое вы хотели бы добавить',
-	'vet-warn3' => 'Приведите, пожалуйста, название видео, которое вы хотели бы добавить',
-	'vet-imagelink' => '<b>Новинка!</b>Добавление видео',
-	'vet-imagebutton' => 'Вставить видео',
-	'vet-license-cc' => 'Лицензировано на условиях Creative Commons Attribution-Share Alike 3.0',
-	'vet-name-incorrect' => 'Это название видео содержит недопустимые символы, как #',
-	'vet-description' => 'Введите полный URL любого из поддерживаемых сайтов. Смотреть',
-	'vet-uploadtext' => "'''Эта форма позволяет быстро загружать и отображать файлы в статьях.'''
-
-''См. также [[Special:Log/upload|журнал загрузок]] и [[Special:ImageList|список всех загруженных файлов]].''",
-	'vet-show-message' => 'показать сообщение',
-	'vet-hide-message' => 'скрыть сообщение',
-	'vet-hide-license-msg' => 'скрыть текст лицензии',
-	'vet-show-license-msg' => 'показать текст лицензии',
-	'vet-max-thumb' => 'превышен максимальный размер миниатюры, возвращение к исходному размеру',
-	'vet-notallowed' => 'Вы не можете вставлять видео в этой вики.',
-	'vet-title' => 'Добавить видео на вашу статью',
-	'vet-supported' => 'Поддерживаемые видео-сайты:',
-	'vet-preview' => '<i>Данный предварительный просмотр не в масштабе</i>',
-	'vet-bad-url' => 'Предоставленный URL является недействительным',
-	'vet-non-existing' => 'Предоставленного видео не существует',
-	'vet-main-info' => '<i>Введите полный URL любой из поддерживаемых сайтов</i>',
-	'vet-page-success' => 'Страница видео была успешно обновлена. Возможно, вам придется сбросить кэш страницы, чтобы увидеть изменения.',
-	'vet-add-button' => 'Добавить видео',
-	'vet-added-from-gallery' => 'видео добавлено из видеогалереи',
-	'vet-see-all' => 'всё',
-	'vet-bad-search' => 'Ошибка при загрузке поискового запроса.',
-	'vet-no-preview' => 'Вы не можете добавить видео из статьи в режиме предпросмотра',
-	'vet-single-success' => 'Видео было успешно размещено в этой статье',
-	'vet-gallery-create-success' => 'Видеогалерея была добавлена к этой статье',
-	'vet-gallery-add-success' => 'Видео было добавлено в видеогалерею этой статьи',
-	'vet-left' => 'Слева',
-	'vet-right' => 'Справа',
-	'vet-gallery' => 'Галерея',
-	'vet-protected' => 'Это видео защищено',
-);
-
-/** Slovenian (slovenščina)
- * @author Dbc334
- */
-$messages['sl'] = array(
-	'vet-back' => 'Nazaj',
-);
-
-/** Serbian (Cyrillic script) (‪српски (ћирилица)‬)
- * @author Rancher
- * @author Verlor
- */
-$messages['sr-ec'] = array(
-	'vet-upload' => 'Пуна URL адреса',
-	'vet-notlogged' => 'Пријавите се или отворите налог да бисте могли да убацујете видео-снимке',
-	'vet-find' => 'Пронађи видео-снимак',
-	'vet-thiswiki' => 'Овај вики',
-	'vet-flickr' => 'Фликр',
-	'vet-upload-btn' => 'Следеће',
-	'vet-find-btn' => 'Претражи',
-	'vet-conflict-inf' => '<h1>Већ постоји видео-снимак с тим именом.</h1>Шта желите да урадите са <b>$1</b>?',
-	'vet-rename' => 'Преименуј видео-снимак',
-	'vet-existing' => 'Користи постојећи видео-снимак',
-	'vet-overwrite' => 'Замени постојећи видео-снимак с мојим',
-	'vet-insert' => 'Убаци',
-	'vet-details-inf' => '<h1>Подаци о датотеци</h1>Наведите име видео-снимка. Треба да буде описно.',
-	'vet-name' => 'Назив',
-	'vet-license' => 'Лиценца',
-	'vet-details-inf2' => 'Поставке приказа видео-снимка',
-	'vet-size' => 'Величина',
-	'vet-thumbnail' => 'Минијатура',
-	'vet-fullsize' => 'Пуна величина ($1×$2 тачака)',
-	'vet-width' => 'Ширина',
-	'vet-layout' => 'Распоред',
-	'vet-caption' => 'Назив',
-	'vet-optional' => '(необавезно)',
-	'vet-insert2' => 'Убаци видео-снимак',
-	'vet-update' => 'Ажурирај видео-снимак',
-	'vet-video-replace-link' => 'Да бисте заменили овај видео-снимак, идите на [[:$1]]',
-	'vet-prev' => 'Претходно',
-	'vet-next' => 'Следеће',
-	'vet-metacafe' => 'Видео-снимци на Метакафију ($1 {{PLURAL:$1|резултат|резултата|резултата}})',
-	'vet-insert3' => 'Убаци видео-снимак',
-	'vet-recent-inf' => 'Недавно убачени видео-снимци',
-	'vet-thiswiki2' => 'Видео-снимци на овом викију ($1 {{PLURAL:$1|резултат|резултата|резултата}})',
-	'vet-success' => '<h1>Успешно извршено</h1>',
-	'vet-return' => 'Назад на уређивање',
-	'vet-back' => 'Назад',
-	'vet-close' => 'Затвори',
-	'vet-imagelink' => '<b>Ново!</b> Убацивање видео-снимака',
-	'vet-imagebutton' => 'Убаци видео-снимке',
-	'vet-show-message' => 'прикажи поруку',
-	'vet-hide-message' => 'сакриј поруку',
-	'vet-hide-license-msg' => 'сакриј текст лиценце',
-	'vet-show-license-msg' => 'прикажи текст лиценце',
-	'vet-notallowed' => 'Не можете да убацујете видео-снимке на овај вики',
-	'vet-title' => 'Додавање видео-снимка у чланак',
-	'vet-supported' => 'Подржани видео-сајтови:',
-	'vet-bad-url' => 'URL адреса је неисправна',
-	'vet-non-existing' => 'Наведени видео-снимак не постоји',
-	'vet-add-button' => 'Додај видео-снимак',
-	'vet-see-all' => 'све',
-	'vet-left' => 'Лево',
-	'vet-right' => 'Десно',
-	'vet-gallery' => 'Галерија',
-	'vet-protected' => 'Овај видео-снимак је заштићен.',
-);
-
-/** Swedish (svenska)
- * @author Jopparn
- * @author Lokal Profil
- * @author Tobulos1
- * @author WikiPhoenix
- */
-$messages['sv'] = array(
-	'vet-upload' => 'Fullständig URL',
-	'vet-notlogged' => 'Logga in eller registrera dig för att lägga till videoklipp',
-	'vet-find' => 'Hitta en video',
-	'vet-thiswiki' => 'Denna wiki',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Nästa',
-	'vet-find-btn' => 'Sök',
-	'vet-flickr-inf' => 'Sök ordentligt licensierade fria foton från Flickr och importera dem till din wiki.',
-	'vet-conflict-inf' => '<h1>Hoppsan, en video med det namnet finns redan.</h1>Vad tycker du om <b>$1</b>?',
-	'vet-rename' => 'Byt namn på video',
-	'vet-existing' => 'Använd befintlig video',
-	'vet-overwrite' => 'Skriv över den befintliga video med din',
-	'vet-insert' => 'Infoga',
-	'vet-details-inf' => '<h1>Fildetaljer</h1>Ange ett namn för den här videon. Gör den något mer beskrivande.',
-	'vet-name' => 'Namn',
-	'vet-license' => 'Licens',
-	'vet-details-inf2' => 'Visningsalternativ för video',
-	'vet-size' => 'Storlek',
-	'vet-thumbnail' => 'Miniatyrbild',
-	'vet-fullsize' => 'Full storlek ($1x$2px)',
-	'vet-width' => 'Bredd',
-	'vet-layout' => 'Layout',
-	'vet-caption' => 'Bildtext',
-	'vet-optional' => '(Tillval)',
-	'vet-insert2' => 'Infoga video',
-	'vet-update' => 'Uppdatera video',
-	'vet-video-replace-link' => 'För att ersätta den här videon, gå till [[:$1]]',
-	'vet-prev' => 'Föregående 8',
-	'vet-next' => 'Nästa 8',
-	'vet-metacafe' => 'Metacafe videoklipp ($1 {{PLURAL:$1|resultat|resultat}})',
-	'vet-insert3' => 'Infoga den här videon',
-	'vet-recent-inf' => 'Nyligen infogade videoklipp',
-	'vet-thiswiki2' => 'Denna wikins videoklipp ($1 {{PLURAL:$1|resultat|resultat}})',
-	'vet-success' => '<h1>Succé</h1>',
-	'vet-return' => 'Återgå till redigering',
-	'vet-back' => 'Tillbaka',
-	'vet-close' => 'Stäng',
-	'vet-warn1' => 'Du måste ange en fråga först!',
-	'vet-warn2' => 'Vänligen ange URL till videon du vill lägga till',
-	'vet-warn3' => 'Vänligen ange namnet på den video du vill lägga till',
-	'vet-imagelink' => '<b>Nytt!</b> Infoga videoklipp',
-	'vet-imagebutton' => 'Infoga videoklipp',
-	'vet-license-cc' => 'Gjord tillgänglig under licensen Creative Commons Erkännande-Dela Lika 3.0',
-	'vet-name-incorrect' => 'Detta videonamn innehåller ogiltiga tecken, som #',
-	'vet-description' => 'Skriv in den fulla URL-adressen från någon av de stödjande sidorna. Se',
-	'vet-uploadtext' => "'''Detta formulär tillåter dig att snabbt ladda upp och visa filer i artiklar.'''
-
-''Se också [[Special:Log/upload|uppladdnings-loggen]] och en [[Special:ImageList|lista av alla uppladdade filer]].''",
-	'vet-show-message' => 'visa meddelande',
-	'vet-hide-message' => 'dölj meddelande',
-	'vet-hide-license-msg' => 'dölj licens-text',
-	'vet-show-license-msg' => 'visa licens-text',
-	'vet-max-thumb' => 'maximala miniatyrstorlek överskrids, återgår till ursprunglig storlek',
-	'vet-notallowed' => 'Du kan inte infoga videoklipp på denna wiki',
-	'vet-title' => 'Lägg till en video till din artikel',
-	'vet-supported' => 'Stödjande videosidor:',
-	'vet-preview' => '<i>Denna förhandsvisning är inte i skala</i>',
-	'vet-bad-url' => 'Den medföljande webbadressen är ogiltig',
-	'vet-non-existing' => 'Den medföljande videon finns inte',
-	'vet-main-info' => '<i>Skriv in den fulla URL-adressen från någon av de stödjande sidorna</i>',
-	'vet-page-success' => 'Videosida har uppdaterats. Du kanske måste rensa videosidan för att se förändringen.',
-	'vet-add-button' => 'Lägg till video',
-	'vet-added-from-gallery' => 'videon har lagts till från video-galleriet',
-	'vet-see-all' => 'alla',
-	'vet-bad-search' => 'Fel uppstod under inläsning av sökfråga.',
-	'vet-no-preview' => 'Du kan inte lägga till videon från artikeln i förhandsgranskning.',
-	'vet-single-success' => 'Video har placerats i denna artikel.',
-	'vet-gallery-create-success' => 'Ett video-galleri lades till i denna artikel.',
-	'vet-gallery-add-success' => 'Videoklippet har lagts till i video-galleriet på den här artikeln.',
-	'vet-left' => 'Vänster',
-	'vet-right' => 'Höger',
-	'vet-gallery' => 'Galleri',
-	'vet-protected' => 'Denna video är skyddad.',
-);
-
-/** Swahili (Kiswahili)
- * @author Lloffiwr
- */
-$messages['sw'] = array(
-	'vet-left' => 'Kushoto',
-	'vet-right' => 'Kulia',
-);
-
-/** Tamil (தமிழ்)
- * @author Karthi.dr
- */
-$messages['ta'] = array(
-	'vet-thiswiki' => 'இந்த விக்கி',
-	'vet-upload-btn' => 'அடுத்தது',
-	'vet-find-btn' => 'தேடுக',
-	'vet-insert' => 'உள்ளிடு',
-	'vet-name' => 'பெயர்',
-	'vet-license' => 'உரிமம்',
-	'vet-size' => 'அளவு',
-	'vet-thumbnail' => 'நகம் அளவு சிறுபடம்',
-	'vet-width' => 'அகலம்',
-	'vet-layout' => 'வடிவமைப்பு',
-	'vet-next' => 'அடுத்து  8',
-	'vet-back' => 'பின் செல்',
-	'vet-close' => 'மூடுக',
-);
-
-/** Telugu (తెలుగు)
- * @author Veeven
- */
-$messages['te'] = array(
-	'vet-upload-btn' => 'తర్వాతి',
-	'vet-find-btn' => 'వెతుకు',
-	'vet-name' => 'పేరు',
-	'vet-size' => 'పరిమాణం',
-	'vet-width' => 'వెడల్పు',
-	'vet-optional' => '  (ఐచ్చికం)',
-	'vet-back' => 'వెనక్కి',
-	'vet-close' => 'మూసివేయి',
-);
-
-/** Tetum (tetun)
- * @author MF-Warburg
- */
-$messages['tet'] = array(
-	'vet-name' => 'Naran',
-);
-
-/** Thai (ไทย)
- * @author Saipetch
- */
-$messages['th'] = array(
-	'vet-find' => 'ค้นหาวิดีโอ',
-	'vet-find-btn' => 'ค้นหา',
-	'vet-rename' => 'เปลี่ยนชื่อวิดีโอ',
-	'vet-name' => 'ชื่อ',
-	'vet-size' => 'ขนาด',
-	'vet-width' => 'ความกว้าง',
-	'vet-close' => 'ปิด',
-	'vet-warn3' => 'กรุณาใส่ชื่อของวิดีโอที่คุณต้องการเพิ่ม',
-	'vet-show-message' => 'แสดงข้อความ',
-	'vet-hide-message' => 'ซ่อนข้อความ',
-	'vet-see-all' => 'ทั้งหมด',
-);
-
-/** Tagalog (Tagalog)
- * @author AnakngAraw
- */
-$messages['tl'] = array(
-	'vet-upload' => 'Buong URL',
-	'vet-notlogged' => 'Lumagda o magpatala upang makapagsingit ng mga bidyo',
-	'vet-find' => 'Maghanap ng isang bidyo',
-	'vet-thiswiki' => 'Ang wiking ito',
-	'vet-flickr' => 'Flickr',
-	'vet-upload-btn' => 'Ikargang paitaas',
-	'vet-find-btn' => 'Maghanap',
-	'vet-flickr-inf' => 'Maghanap ng malalayang mga larawang may marapat na lisensiya mula sa Flickr at angkatin sila papunta sa wiki mo.',
-	'vet-conflict-inf' => '<h1>Naku, umiiral na ang isang bidyong may ganyang pangalan.</h1>Ano ang nais mong gawin tungkol sa <b>$1</b>?',
-	'vet-rename' => 'Muling pangalanan ang bidyo',
-	'vet-existing' => 'Gamitin ang umiiral na bidyo',
-	'vet-overwrite' => 'Patungan ng iyo ang umiiral na bidyo',
-	'vet-insert' => 'Isingit',
-	'vet-details-inf' => '<h1>Mga detalye ng talaksan</h1>Tumukoy ng isang pangalan para sa bidyong ito.  Gawin itong isang bagay na mapaglarawan.',
-	'vet-name' => 'Pangalan',
-	'vet-license' => 'Lisensiya',
-	'vet-details-inf2' => 'Mga mapagpipilian ng pagpapakita ng bidyo',
-	'vet-size' => 'Sukat',
-	'vet-thumbnail' => 'Kagyat',
-	'vet-fullsize' => 'Buong-sukat ($1x$2px)',
-	'vet-width' => 'Lapad',
-	'vet-layout' => 'Kalatagan',
-	'vet-caption' => 'Paliwanag',
-	'vet-optional' => '(Maaaring wala nito)',
-	'vet-insert2' => 'Isingit ang bidyo',
-	'vet-update' => 'Isapanahon ang bidyo',
-	'vet-video-replace-link' => 'Upang mapalitan ang bidyong ito, mangyaring pumunta sa [[:$1]]',
-	'vet-prev' => 'Nakaraang 8',
-	'vet-next' => 'Susunod na 8',
-	'vet-metacafe' => 'Mga bidyo ng Kapihang Meta ($1 {{PLURAL:$1|kinalabasan|mga kinalabasan}})',
-	'vet-insert3' => 'Isingit ang bidyong ito',
-	'vet-recent-inf' => 'Kamakailang isingit na mga bidyo',
-	'vet-thiswiki2' => 'Ang bidyo ng wiking ito ($1 {{PLURAL:$1|kinalabasan|mga kinalabasan}})',
-	'vet-success' => '<h1>Tagumpay</h1>',
-	'vet-return' => 'Magbalik sa pamamatnugot',
-	'vet-back' => 'Bumalik',
-	'vet-close' => 'Isara',
-	'vet-warn1' => 'Kailangan mo munang tukuyin ang tanong!',
-	'vet-warn2' => 'Mangyaring ibigay ang URL ng bidyong nais mong idagdag',
-	'vet-warn3' => 'Mangyaring ibigay ang pangalan ng bidyong nais mong idagdag',
-	'vet-imagelink' => '<b>Bago!</b> Magsingit ng mga bidyo',
-	'vet-imagebutton' => 'Isingit ang mga bidyo',
-	'vet-license-cc' => 'Nakalisensya sa ilalim ng Creative Commons Attribution-Share Alike 3.0',
-	'vet-name-incorrect' => 'Ang pangalan bidyong ito ay naglalaman ng hindi katanggap-tanggap na mga panitik',
-	'vet-description' => 'Ipasok ang buong URL mula sa anuman sa tinatangkilik na mga sityo. Tingnan ang',
-	'vet-show-message' => 'ipakita ang mensahe',
-	'vet-hide-message' => 'itago ang mensahe',
-	'vet-hide-license-msg' => 'itago ang teksto ng lisensya',
-	'vet-show-license-msg' => 'ipakita ang teksto ng lisensya',
-	'vet-max-thumb' => 'lumabis sa pinakamataas na sukat ng kagyat, nagpapanumbalik sa orihinal na sukat',
-	'vet-notallowed' => 'Hindi ka makapagsisingit ng mga bidyo sa wiking ito',
-	'vet-title' => 'Magdagdag ng isang bidyo sa artikulo mo',
-	'vet-supported' => 'Tinatangkilik na mga sityo ng bidyo:',
-	'vet-preview' => '<i>Wala sa sukat ang paunang tinging ito</i>',
-	'vet-bad-url' => 'Hindi katanggap-katanggap ang ibinigay na URL',
-	'vet-non-existing' => 'Hindi umiiral ang ibinigay na bidyo',
-	'vet-main-info' => '<i>Ipasok ang buong URL mula sa anuman sa tinatangkilik na mga sityo</i>',
-	'vet-page-success' => 'Matagumpay na naisapanahon ang pahina ng bidyo. Maaaring kailanganin mong purgahin ang pahina ng bidyo upang makita ang pagbabago.',
-	'vet-add-button' => 'Idagdag ang bidyo',
-	'vet-added-from-gallery' => 'idinagdag ang bidyong nagmula sa galeriya ng bidyo',
-	'vet-see-all' => 'lahat',
-	'vet-bad-search' => 'Kamalian sa pagkakarga ng tanong ng paghanap.',
-	'vet-no-preview' => 'Hindi ka maaaring magdagdag ng bidyo mula sa artikulo habang nasa pamamaraan ng paunang tingin.',
-	'vet-single-success' => 'Matagumpay na nailagay ang bidyo sa loob ng artikulong ito.',
-	'vet-gallery-create-success' => 'Isang tanghalan ng bidyo ang naidagdag sa artikulong ito.',
-	'vet-gallery-add-success' => 'Isang bidyo ang naidagdag sa tanghalan ng bidyo sa loob ng artikulong ito.',
-	'vet-left' => 'Kaliwa',
-	'vet-right' => 'Kanan',
-	'vet-gallery' => 'Galeriya',
-	'vet-protected' => 'Pinuprutektahan ang bidyong ito.',
-);
-
-/** Ukrainian (українська)
- * @author Prima klasy4na
- * @author Тест
- */
-$messages['uk'] = array(
-	'vet-find-btn' => 'Знайти',
-	'vet-insert' => 'Вставити',
-	'vet-name' => 'Назва',
-	'vet-license' => 'Ліцензія',
-	'vet-size' => 'Розмір',
-	'vet-thumbnail' => 'Мініатюра',
-	'vet-prev' => 'Попередні 8',
-	'vet-next' => 'Наступні 8',
-	'vet-back' => 'Назад',
-	'vet-close' => 'Закрити',
-);
-
-/** Yiddish (ייִדיש)
- * @author פוילישער
- */
-$messages['yi'] = array(
-	'vet-license' => 'ליצענץ',
-	'vet-caption' => 'באשרייבונג',
-);
-
-/** Simplified Chinese (‪中文（简体）‬)
- * @author Anakmalaysia
- * @author Hydra
- * @author Xiaomingyan
- */
-$messages['zh-hans'] = array(
-	'vet-thiswiki' => '这个维基',
-	'vet-upload-btn' => '下一步',
-	'vet-find-btn' => '寻找',
-	'vet-insert' => '插入',
-	'vet-name' => '名',
-	'vet-license' => '授权协议',
-	'vet-size' => '大小',
-	'vet-prev' => '前 8 个',
-	'vet-next' => '后 8 个',
-	'vet-back' => '后退',
-	'vet-close' => '关闭',
-	'vet-show-message' => '显示性息',
-	'vet-see-all' => '全部',
-	'vet-left' => '左',
-	'vet-right' => '右',
-);
-
