--- extensions/wikia/VideoEmbedTool2/WikiaVideoAdd.i18n.php
@@ -1,812 +1,0 @@
-<?php
-
-$messages = array();
-
-$messages['en'] = array(
-	'wva-success' => 'Video page $1 was successfully added.',
-	'wva-success-replaced' => 'Video page $1 was successfully replaced.',
-	'wva-failure' => 'Please supply valid parameters.',
-	'wva-name' => 'Enter the name:',
-	'wva-url' => 'Enter the URL:',
-	'wva-add' => 'Add',
-	'wva-title' => 'Quickly add video From URL',
-	'wva-notlogged' => 'You need to log in or register to insert videos.',
-	'wva-notallowed' => 'You cannot insert videos on this wiki.',
-	'wva-protected' => 'This video is protected.',
-	'wva-thumbnail-upload-failed' => 'Video embed failed, unable to get proper thumbnail'
-);
-
-/** Message documentation (Message documentation)
- * @author EugeneZelenko
- */
-$messages['qqq'] = array(
-	'wva-add' => '{{Identical|Add}}'
-);
-
-/** Afrikaans (Afrikaans)
- * @author Naudefj
- */
-$messages['af'] = array(
-	'wva-success' => 'Videobladsy $1 was suksesvol bygevoeg.',
-	'wva-success-replaced' => 'Videobladsy $1 was suksesvol vervang.',
-	'wva-failure' => 'Verskaf asseblief geldige parameters.',
-	'wva-name' => 'Sleutel die naam in:',
-	'wva-url' => 'Sleutel die URL in:',
-	'wva-add' => 'Voeg by',
-	'wva-title' => "Voeg vinnig 'n video vanaf URL by",
-	'wva-notlogged' => "U moet aanteken of registreer om video's by te kan voeg.",
-	'wva-notallowed' => "U kan nie video's op hierdie wiki byvoeg nie.",
-);
-
-/** Arabic (العربية)
- * @author Achraf94
- * @author Aiman titi
- */
-$messages['ar'] = array(
-	'wva-success' => 'صفحة الفيديو $1 تم إضافتها بنجاح',
-	'wva-success-replaced' => 'صفحة الفيديو $1 تم تعويضها بنجاح',
-	'wva-failure' => 'يرجى تزويد المعلومات الصحيحة',
-	'wva-name' => 'أدخل الإسم:',
-	'wva-url' => 'أدخل العنوان الإلكتروني (URL):',
-	'wva-add' => 'أضف',
-	'wva-title' => 'إضافة فيديو بسرعة من العنوان (URL)',
-	'wva-notlogged' => 'يجب أن تكون مسجلا إذا أرد إضافة فيديو',
-	'wva-notallowed' => 'لا يمكنك إضافة فيديو في هذه الويكي',
-	'wva-protected' => 'هذا الفيديو. محمي',
-);
-
-/** Azerbaijani (azərbaycanca)
- * @author Cekli829
- */
-$messages['az'] = array(
-	'wva-add' => 'Əlavə et',
-);
-
-/** Belarusian (Taraškievica orthography) (‪беларуская (тарашкевіца)‬)
- * @author EugeneZelenko
- * @author Jim-by
- */
-$messages['be-tarask'] = array(
-	'wva-success' => 'Старонка відэа $1 пасьпяхова дададзеная.',
-	'wva-success-replaced' => 'Старонка відэа $1 пасьпяхова замененая.',
-	'wva-failure' => 'Калі ласка, падайце слушныя парамэтры.',
-	'wva-name' => 'Увядзіце назву:',
-	'wva-url' => 'Увядзіце URL-адрас:',
-	'wva-add' => 'Дадаць',
-	'wva-title' => 'Хуткае даданьне відэа па URL-адрасу',
-	'wva-notlogged' => 'Каб дадаць відэа, Вам неабходна ўвайсьці ў сыстэму альбо зарэгістравацца.',
-	'wva-notallowed' => 'Вы ня можаце устаўляць відэа ў {{GRAMMAR:месны|{{SITENAME}}}}.',
-	'wva-protected' => 'Гэтае відэа абароненае.',
-);
-
-/** Bulgarian (български)
- * @author DCLXVI
- */
-$messages['bg'] = array(
-	'wva-add' => 'Добавяне',
-);
-
-/** Banjar (Bahasa Banjar)
- * @author Ezagren
- */
-$messages['bjn'] = array(
-	'wva-add' => 'Tambahakan',
-);
-
-/** Breton (brezhoneg)
- * @author Fulup
- * @author Gwenn-Ael
- * @author Y-M D
- */
-$messages['br'] = array(
-	'wva-success' => 'Ar bajenn video $1 zo bet ouzhpennet gant berzh',
-	'wva-success-replaced' => "Ar bajenn video $1 zo bet erlerc'hiet gant berzh",
-	'wva-failure' => 'Ouzhpennañ arventennoù reizh, mar plij',
-	'wva-name' => 'Lakait an anv :',
-	'wva-url' => 'Ebarzhiñ an URL :',
-	'wva-add' => 'Ouzhpennañ',
-	'wva-title' => 'Ouzhpennañ buan ur video adalek e URL',
-	'wva-notlogged' => "Ezhomm ho peus da vezañ kevreañ pe enrollet evit ensoc'hañ videoioù.",
-	'wva-notallowed' => "Ne c'hallit ket ensoc'hañ videoioù war ar wiki-mañ",
-	'wva-protected' => 'Gwarezet eo ar video-mañ',
-);
-
-/** Catalan (català)
- * @author Gemmaa
- * @author Paucabot
- */
-$messages['ca'] = array(
-	'wva-success' => "Pàgina de vídeo de  $1  s'afegia reeixidament.",
-	'wva-success-replaced' => 'Pàgina de vídeo de  $1  va ser substituït amb èxit.',
-	'wva-failure' => 'Si us plau proporcionar els paràmetres vàlids.',
-	'wva-name' => 'Escriviu el nom:',
-	'wva-url' => 'Introduïu la URL:',
-	'wva-add' => 'Afegeix',
-	'wva-title' => "De pressa afegir des d'URL de vídeo",
-	'wva-notlogged' => "Vostè necessita connectar-se o inscriure's per inserir vídeos.",
-	'wva-notallowed' => 'No es pot inserir vídeos en aquest wiki.',
-	'wva-protected' => 'Aquest vídeo està protegit.',
-);
-
-/** Czech (česky)
- * @author Jezevec
- * @author Mormegil
- * @author Mr. Richard Bolla
- */
-$messages['cs'] = array(
-	'wva-success' => 'Stránka videa $1 byla úspěšně přidána.',
-	'wva-success-replaced' => 'Stránka videa $1 byla úspěšně nahrazena.',
-	'wva-failure' => 'Uveďte prosím platné parametry.',
-	'wva-name' => 'Zadejte název:',
-	'wva-url' => 'Zadejte URL:',
-	'wva-add' => 'Přidat',
-	'wva-title' => 'Přidat rychle video z URL',
-	'wva-notlogged' => 'Pro vkládání videí se musíte přihlásit nebo zaregistrovat.',
-	'wva-notallowed' => 'Na této wiki nemůžete vkládat videa.',
-	'wva-protected' => 'Toto video je zamčeno.',
-	'wva-thumbnail-upload-failed' => 'Vložení videa selhalo, nelze získat odpovídající miniaturu',
-);
-
-/** German (Deutsch)
- * @author LWChris
- * @author PtM
- * @author The Evil IP address
- */
-$messages['de'] = array(
-	'wva-success' => 'Videoseite für $1 wurde erfolgreich erstellt.',
-	'wva-success-replaced' => 'Videoseite für $1 wurde erfolgreich ersetzt.',
-	'wva-failure' => 'Gib bitte gültige Parameter ein.',
-	'wva-name' => 'Name eingeben:',
-	'wva-url' => 'Gib die URL hier ein:',
-	'wva-add' => 'Hinzufügen',
-	'wva-title' => 'Video von URL hinzufügen',
-	'wva-notlogged' => 'Du musst dich einloggen oder registrieren um Videos einzufügen.',
-	'wva-notallowed' => 'Du kannst in diesem Wiki keine Videos einfügen.',
-	'wva-protected' => 'Dieses Video ist geschützt',
-	'wva-thumbnail-upload-failed' => 'Videoeinbettung fehlgeschlagen, kein brauchbares Vorschaubild erhalten',
-);
-
-/** German (formal address) (‪Deutsch (Sie-Form)‬)
- * @author LWChris
- */
-$messages['de-formal'] = array(
-	'wva-success' => 'Videoseite für $1 wurde erfolgreich erstellt.',
-	'wva-success-replaced' => 'Videoseite für $1 wurde erfolgreich ersetzt.',
-	'wva-failure' => 'Geben Sie bitte gültige Parameter ein.',
-	'wva-name' => 'Name eingeben:',
-	'wva-url' => 'Geben Sie die URL hier ein:',
-	'wva-add' => 'Hinzufügen',
-	'wva-title' => 'Video von URL hinzufügen',
-	'wva-notlogged' => 'Sie müssen sich einloggen oder registrieren um Videos einzufügen.',
-	'wva-notallowed' => 'Sie können in diesem Wiki keine Videos einfügen.',
-);
-
-/** Zazaki (Zazaki)
- * @author Erdemaslancan
- */
-$messages['diq'] = array(
-	'wva-success' => 'Vidyo pelaya $1  vıraziyaya',
-	'wva-success-replaced' => 'Vidyo pelaya $1  re ca deya.',
-	'wva-name' => 'Name deke:',
-	'wva-url' => 'URL deke:',
-	'wva-add' => 'Deke',
-);
-
-/** Esperanto (Esperanto)
- * @author Yekrats
- */
-$messages['eo'] = array(
-	'wva-add' => 'Aldoni',
-);
-
-/** Spanish (español)
- * @author Danke7
- * @author Translationista
- * @author VegaDark
- */
-$messages['es'] = array(
-	'wva-success' => 'La página del vídeo $1 fue añadida satisfactoriamente.',
-	'wva-success-replaced' => 'La página del vídeo $1 fue reemplazada satisfactoriamente.',
-	'wva-failure' => 'Por favor, pon parámetros válidos.',
-	'wva-name' => 'Introduce el nombre:',
-	'wva-url' => 'Introducir la URL:',
-	'wva-add' => 'Añadir',
-	'wva-title' => 'Añadir vídeo de forma rápida desde una URL',
-	'wva-notlogged' => 'Necesitas acceder o registrarte para insertar videos.',
-	'wva-notallowed' => 'No puedes insertar videos en este wiki.',
-	'wva-protected' => 'Este video está protegido',
-	'wva-thumbnail-upload-failed' => 'Error al añadir el vídeo, no se pudo obtener una miniatura adecuada',
-);
-
-/** Basque (euskara)
- * @author An13sa
- */
-$messages['eu'] = array(
-	'wva-add' => 'Gehitu',
-);
-
-/** Persian (فارسی)
- * @author Ebraminio
- * @author Huji
- * @author ZxxZxxZ
- */
-$messages['fa'] = array(
-	'wva-success' => 'فیلم صفحهٔ $1 موفقیت‌آمیز اضافه شد.',
-	'wva-success-replaced' => 'فیلم صفحهٔ $1 موفقیت‌آمیز جایگزین شد.',
-	'wva-failure' => 'لطفا ورودی‌های درستی وارد کنید.',
-	'wva-name' => 'نام را وارد کنید:',
-	'wva-url' => 'نشانی اینترنتی را وارد کنید:',
-	'wva-add' => 'افزودن',
-	'wva-title' => 'به سرعت فیلم را از نشانی اینترنتی واردکردن',
-	'wva-notlogged' => 'برای وارد کردن فیلم، شما باید به سامانه وارد شوید یا ثبت نام کنید.',
-	'wva-notallowed' => 'شما نمی‌توانید در این ویکی فیلم وارد کنید.',
-	'wva-protected' => 'این ویدیو محافظت شده است.',
-);
-
-/** Finnish (suomi)
- * @author Centerlink
- * @author Crt
- * @author Silvonen
- * @author Tofu II
- */
-$messages['fi'] = array(
-	'wva-success' => 'Videosivun $1 lisääminen onnistui.',
-	'wva-success-replaced' => 'Videosivun $1 korvaaminen onnistui.',
-	'wva-failure' => 'Toimita kelvolliset parametrit.',
-	'wva-name' => 'Kirjoita nimi:',
-	'wva-url' => 'Anna verkko-osoite:',
-	'wva-add' => 'Lisää',
-	'wva-title' => 'Lisää nopeasti video verkko-osoitteesta',
-	'wva-notlogged' => 'Sinun täytyy kirjautua sisään tai rekisteröityä, jotta voit lisätä videoita.',
-	'wva-notallowed' => 'Et voi lisätä videoita tälle wikisivulle.',
-	'wva-protected' => 'Tämä video on suojattu.',
-);
-
-/** French (français)
- * @author Gomoko
- * @author Iluvalar
- */
-$messages['fr'] = array(
-	'wva-success' => 'La page de la vidéo $1 a été ajoutée avec succès.',
-	'wva-success-replaced' => 'La page de vidéo $1 a été remplacée avec succès',
-	'wva-failure' => 'Veuillez fournir des paramètres valides',
-	'wva-name' => 'Entrez le nom:',
-	'wva-url' => "Entrez l'URL:",
-	'wva-add' => 'Ajouter',
-	'wva-title' => 'Ajouter une vidéo rapidement à partir de son URL',
-	'wva-notlogged' => 'Vous devez vous inscrire ou bien vous enregistrer pour ajouter des vidéos',
-	'wva-notallowed' => 'Vous ne pouvez pas ajouter de vidéos sur ce wiki',
-	'wva-protected' => 'Cette vidéo est protégée',
-	'wva-thumbnail-upload-failed' => 'La vidéo incluse a échouée, impossible de trouver la bonne miniature',
-);
-
-/** Galician (galego)
- * @author Toliño
- */
-$messages['gl'] = array(
-	'wva-success' => 'Engadiuse con éxito a páxina do vídeo $1.',
-	'wva-success-replaced' => 'Substituíuse con éxito a páxina do vídeo $1.',
-	'wva-failure' => 'Por favor, dea parámetros válidos.',
-	'wva-name' => 'Introduza o nome:',
-	'wva-url' => 'Introduza o URL:',
-	'wva-add' => 'Engadir',
-	'wva-title' => 'Engadir un vídeo de xeito rápido desde un URL',
-	'wva-notlogged' => 'Debe acceder ao sistema ou rexistrarse para inserir vídeos.',
-	'wva-notallowed' => 'Non pode inserir vídeos neste wiki.',
-	'wva-protected' => 'Este vídeo está protexido.',
-	'wva-thumbnail-upload-failed' => 'Erro ao engadir o vídeo; non se puido obter a miniatura axeitada',
-);
-
-/** Hebrew (עברית)
- * @author 0ftal
- * @author שומבלע
- */
-$messages['he'] = array(
-	'wva-success' => 'דף הסרטונים $1 נוסף בהצלחה.',
-	'wva-success-replaced' => 'דף הסרטונים $1 הוחלף בהצלחה.',
-	'wva-failure' => 'נא לספק פרמטרים חוקיים.',
-	'wva-name' => 'הכנסת השם:',
-	'wva-url' => 'הוספת כתובת האתר:',
-	'wva-add' => 'הוספה',
-	'wva-title' => 'הוספה מהירה של סרטון מכתובת',
-	'wva-notlogged' => 'התחברו או הרשמו כדי להוסיף סרטונים.',
-	'wva-notallowed' => 'לא ניתן להוסיף סרטונים לאתר.',
-	'wva-protected' => 'סרטון זה מוגן.',
-);
-
-/** Hungarian (magyar)
- * @author Glanthor Reviol
- * @author TK-999
- */
-$messages['hu'] = array(
-	'wva-success' => 'A(z) $1 videóoldalt sikeresen létrehoztad.',
-	'wva-success-replaced' => 'A(z) $1 videóoldalt sikeresen lecserélted.',
-	'wva-failure' => 'Kérlek, érvényes paramétereket adj meg.',
-	'wva-name' => 'Add meg a nevet:',
-	'wva-url' => 'Add meg az URL-t:',
-	'wva-add' => 'Hozzáadás',
-	'wva-title' => 'Videó gyors hozzáadása URL alapján',
-	'wva-notlogged' => 'Lépj be vagy regisztrálj a videók beszúrásához.',
-	'wva-notallowed' => 'Nem szúrhatsz be videókat ezen a wikin.',
-	'wva-protected' => 'Ez a videó védett.',
-	'wva-thumbnail-upload-failed' => 'Nem sikerült a videó beágyazása&mdash;a megfelelő miniatűr előállítása nem volt lehetséges.',
-);
-
-/** Interlingua (interlingua)
- * @author McDutchie
- */
-$messages['ia'] = array(
-	'wva-success' => 'Le pagina de video $1 ha essite addite con successo.',
-	'wva-success-replaced' => 'Le pagina de video $1 ha essite reimplaciate con successo.',
-	'wva-failure' => 'Per favor forni parametros valide.',
-	'wva-name' => 'Entra le nomine:',
-	'wva-url' => 'Entra le URL:',
-	'wva-add' => 'Adder',
-	'wva-title' => 'Adder video rapidemente con URL',
-	'wva-notlogged' => 'Tu debe aperir un session o registrar te pro inserer videos.',
-	'wva-notallowed' => 'Non es possibile inserer videos in iste wiki.',
-	'wva-protected' => 'Iste video es protegite.',
-	'wva-thumbnail-upload-failed' => 'Le insertion del video ha fallite. Non esseva possibile obtener un bon miniatura.',
-);
-
-/** Indonesian (Bahasa Indonesia)
- * @author Farras
- * @author Irwangatot
- * @author Kenrick95
- */
-$messages['id'] = array(
-	'wva-success' => 'halaman Video $1 berhasil ditambahkan.',
-	'wva-success-replaced' => 'halaman Video $1 berhasil diganti.',
-	'wva-failure' => 'Harap berikan parameter yang benar.',
-	'wva-name' => 'Masukkan nama:',
-	'wva-url' => 'Masukkan URL:',
-	'wva-add' => 'Tambahkan',
-	'wva-title' => 'Cara cepat menambah Video dari URL',
-	'wva-notlogged' => 'Anda harus login atau daftar untuk memasukkan video.',
-	'wva-notallowed' => 'Anda tidak dapat menyisipkan video di wiki ini.',
-	'wva-protected' => 'Video ini dilindungi.',
-);
-
-/** Igbo (Igbo)
- * @author Ukabia
- */
-$messages['ig'] = array(
-	'wva-add' => 'Tinyé',
-);
-
-/** Italian (italiano)
- * @author HalphaZ
- * @author Pietrodn
- */
-$messages['it'] = array(
-	'wva-success' => 'La pagina video $1 è stata aggiunta con successo.',
-	'wva-success-replaced' => 'La pagina video $1 è stata sostituita con successo.',
-	'wva-failure' => 'Prego inserisci parametri validi.',
-	'wva-name' => 'Inserisci il nome:',
-	'wva-url' => "Inserisci l'URL:",
-	'wva-add' => 'Aggiungi',
-	'wva-title' => 'Aggiungi rapidamente video da URL',
-	'wva-notlogged' => 'Devi accedere o registrati per inserire video.',
-	'wva-notallowed' => 'Non puoi inserire video su questo wiki.',
-	'wva-protected' => 'Questo video è protetto.',
-);
-
-/** Japanese (日本語)
- * @author Tommy6
- */
-$messages['ja'] = array(
-	'wva-success' => '動画"$1"の追加に成功しました。',
-	'wva-success-replaced' => '動画"$1"の上書きに成功しました。',
-	'wva-failure' => '適切なパラメータを入力してください。',
-	'wva-name' => '動画名:',
-	'wva-url' => '動画のURL:',
-	'wva-add' => '追加',
-	'wva-title' => '手軽にURLから動画を追加',
-	'wva-notlogged' => '動画を追加するにはログインするかアカウントを作成する必要があります。',
-	'wva-notallowed' => 'このウィキでは動画を追加できません。',
-	'wva-protected' => 'この動画は保護されています。',
-	'wva-thumbnail-upload-failed' => '動画の埋め込みに失敗しました。適切なサムネイルを取得できません。',
-);
-
-/** Khmer (ភាសាខ្មែរ)
- * @author គីមស៊្រុន
- */
-$messages['km'] = array(
-	'wva-success' => 'បានបន្ថែមទំព័រវីដេអូ $1 ជាស្ថាពរ',
-	'wva-success-replaced' => 'បានជំនួសទំព័រវីដេអូ $1 ជាស្ថាពរ',
-	'wva-failure' => 'សូមផ្ដល់ប៉ារ៉ាម៉ែតត្រឹមត្រូវ',
-	'wva-name' => 'បញ្ចូលឈ្មោះ៖',
-	'wva-url' => 'បញ្ចូល URL៖',
-	'wva-add' => 'បន្ថែម',
-	'wva-title' => 'បន្ថែមវីដេអូពីURLតាមរបៀបរហ័ស',
-	'wva-notlogged' => 'អ្នកត្រូវតែកត់ឈ្មោះចូលឬចុះឈ្មោះថ្មីដើម្បីមានសិទ្ធិបញ្ចូលវីដេអូ។',
-	'wva-notallowed' => 'អ្នកមិនអាចបញ្ចូលវីដេអូនៅក្នុងវិគីនេះទេ។',
-);
-
-/** Kannada (ಕನ್ನಡ)
- * @author Nayvik
- */
-$messages['kn'] = array(
-	'wva-add' => 'ಸೇರಿಸು',
-);
-
-/** Karachay-Balkar (къарачай-малкъар)
- * @author Къарачайлы
- */
-$messages['krc'] = array(
-	'wva-success' => '$1 видео бет джетишимли къошулгъанды.',
-	'wva-success-replaced' => '$1 видео бет джетишимли ауушдурулгъанды.',
-	'wva-failure' => 'Тилейбиз, джараулу параметрлени кёргюзтюгюз.',
-	'wva-name' => 'Атын джазыгъыз:',
-	'wva-url' => 'URL адресни джазыгъыз:',
-	'wva-add' => 'Къош',
-	'wva-title' => 'URL-ден видеоны дженгил къошууу',
-	'wva-notlogged' => 'Видеоны къошар ючюн системагъа кесигизни танытыгъыз неда регистрацияны ётюгюз.',
-	'wva-notallowed' => 'Бу викиде видео салаллыкъ тюлсюз.',
-);
-
-/** Luxembourgish (Lëtzebuergesch)
- * @author Robby
- */
-$messages['lb'] = array(
-	'wva-name' => 'Gitt w.e.g. den Numm un:',
-	'wva-url' => "Gitt w.e.g. d'URL an:",
-	'wva-add' => 'Derbäisetzen',
-);
-
-/** Macedonian (македонски)
- * @author Bjankuloski06
- */
-$messages['mk'] = array(
-	'wva-success' => 'Видео-страницата $1 е успешно додадена.',
-	'wva-success-replaced' => 'Видео-страницата $1 е успешно заменета.',
-	'wva-failure' => 'Наведете важечки параметри.',
-	'wva-name' => 'Внесете го името:',
-	'wva-url' => 'Внесете URL-адреса:',
-	'wva-add' => 'Додај',
-	'wva-title' => 'Брзо додавање на видеоснимка од URL-адреса',
-	'wva-notlogged' => 'Треба да сте најавени или да се регистрирате за да можете да вметнувате видеоснимки.',
-	'wva-notallowed' => 'Не можете да вметнувате видеоснимки на ова вики.',
-	'wva-protected' => 'Оваа видеоснимка е заштитена.',
-	'wva-thumbnail-upload-failed' => 'Вметнувањето на видеото не успеа. Не можев да добијам правилна минијатура',
-);
-
-/** Mongolian (монгол)
- * @author Chinneeb
- */
-$messages['mn'] = array(
-	'wva-add' => 'Нэмэх',
-);
-
-/** Malay (Bahasa Melayu)
- * @author Anakmalaysia
- */
-$messages['ms'] = array(
-	'wva-success' => 'Laman video $1 berjaya dibuka.',
-	'wva-success-replaced' => 'Laman video $1 berjaya diganti.',
-	'wva-failure' => 'Sila nyatakan parameter yang sah.',
-	'wva-name' => 'Isikan nama:',
-	'wva-url' => 'Isikan URL:',
-	'wva-add' => 'Tambahkan',
-	'wva-title' => 'Tambahkan video dengan segera dari URL',
-	'wva-notlogged' => 'Anda perlu log masuk atau berdaftar untuk menyisipkan video.',
-	'wva-notallowed' => 'Anda tidak boleh menyisipkan video ke dalam wiki ini.',
-	'wva-protected' => 'Video ini dilindungi.',
-	'wva-thumbnail-upload-failed' => 'Pembenaman video gagal, gambar kenit yang sebetulnya tidak dapat diperoleh',
-);
-
-/** Burmese (မြန်မာဘာသာ)
- * @author Erikoo
- */
-$messages['my'] = array(
-	'wva-add' => 'ပေါင်းထည့်ရန်',
-);
-
-/** Norwegian Bokmål (‪norsk (bokmål)‬)
- * @author Audun
- * @author Nghtwlkr
- */
-$messages['nb'] = array(
-	'wva-success' => 'Videoside $1 ble lagt til.',
-	'wva-success-replaced' => 'Videoside $1 ble erstattet.',
-	'wva-failure' => 'Vennligst oppgi gyldige parametre.',
-	'wva-name' => 'Skriv inn navnet:',
-	'wva-url' => 'Skriv inn internettadressen:',
-	'wva-add' => 'Legg til',
-	'wva-title' => 'Legg til video raskt fra internettadresse',
-	'wva-notlogged' => 'Du må logge inn eller registrere deg for å legge til videoer.',
-	'wva-notallowed' => 'Du kan ikke legge til videoer på denne wikien.',
-	'wva-protected' => 'Denne videoen er beskyttet',
-	'wva-thumbnail-upload-failed' => 'Videoinnebyggingen mislyktes, kunne ikke hente miniatyrbilde',
-);
-
-/** Dutch (Nederlands)
- * @author Siebrand
- */
-$messages['nl'] = array(
-	'wva-success' => 'De videopagina $1 is toegevoegd.',
-	'wva-success-replaced' => 'De videopagina $1 is vervangen.',
-	'wva-failure' => 'Geef alstublieft geldige parameters op.',
-	'wva-name' => 'Geef de naam in:',
-	'wva-url' => 'Geef de URL in:',
-	'wva-add' => 'Toevoegen',
-	'wva-title' => 'Video vanuit URL toevoegen',
-	'wva-notlogged' => "U moet aanmelden of registreren om video's te kunnen toevoegen.",
-	'wva-notallowed' => "U kunt geen video's toevoegen in deze wiki.",
-	'wva-protected' => 'Deze video is beveiligd.',
-	'wva-thumbnail-upload-failed' => 'Het inbedden van de video is mislukt. Het was niet mogelijk om een miniatuurafbeelding te maken',
-);
-
-/** ‪Nederlands (informeel)‬ (‪Nederlands (informeel)‬)
- * @author Siebrand
- */
-$messages['nl-informal'] = array(
-	'wva-notlogged' => "Je moet aanmelden of registreren om video's te kunnen toevoegen.",
-	'wva-notallowed' => "Je kunt geen video's toevoegen in deze wiki.",
-);
-
-/** Occitan (occitan)
- * @author Cedric31
- */
-$messages['oc'] = array(
-	'wva-add' => 'Apondre',
-	'wva-title' => 'Apondre una vidèo rapidament a partir de son URL',
-);
-
-/** Papiamento (Papiamentu)
- * @author Sdm1985
- */
-$messages['pap'] = array(
-	'wva-success' => 'Página di vidio $1 a wòrdu agregá eksitosamente',
-	'wva-success-replaced' => 'Página di video $1 a wòrdu reemplasá eksitosamente',
-	'wva-failure' => 'Por fabor provee parametonan balido',
-	'wva-name' => 'Hinka e nòmber',
-	'wva-url' => 'Hinka e adrès di URL:',
-	'wva-add' => 'Agregá',
-	'wva-title' => 'Agregá vidio via URL di forma rápido',
-	'wva-notlogged' => 'Bo mester login of inskribí bo mes pa por hinka vidionan.',
-	'wva-notallowed' => 'Bo no por hinka vidio den e wiki aki.',
-);
-
-/** Polish (polski)
- * @author Marcin Łukasz Kiejzik
- * @author Sovq
- * @author Sp5uhe
- */
-$messages['pl'] = array(
-	'wva-success' => 'Strona filmu „$1” została pomyślnie dodana.',
-	'wva-success-replaced' => 'Strona filmu „$1” została pomyślnie zastąpiona.',
-	'wva-failure' => 'Podaj prawidłowe parametry.',
-	'wva-name' => 'Wpisz nazwę:',
-	'wva-url' => 'Podaj URL:',
-	'wva-add' => 'Dodaj',
-	'wva-title' => 'Dodaj film z adresu URL',
-	'wva-notlogged' => 'Dodawanie filmów jest możliwe po zarejestrowaniu i zalogowaniu się.',
-	'wva-notallowed' => 'Nie możesz dodawać filmów do tej wiki.',
-	'wva-protected' => 'Ten film jest zabezpieczony.',
-	'wva-thumbnail-upload-failed' => 'Osadzanie filmu nie powiodło się, wystąpił błąd w trakcie generowania miniatury.',
-);
-
-/** Piedmontese (Piemontèis)
- * @author Borichèt
- * @author Dragonòt
- */
-$messages['pms'] = array(
-	'wva-success' => "La pàgina video $1 a l'é stàita giontà da bin.",
-	'wva-success-replaced' => "La pàgina video $1 a l'é stàita rimpiassà da bin.",
-	'wva-failure' => 'Për piasì dà paràmetr bon.',
-	'wva-name' => 'Anserì ël nòm:',
-	'wva-url' => "Anserì l'URL:",
-	'wva-add' => 'Gionta',
-	'wva-title' => 'Gionta video an pressa da URL',
-	'wva-notlogged' => 'A deuv rintré ant ël sistema o argistresse për anserì ëd filmà',
-	'wva-notallowed' => 'It peule pa anserì video su sta wiki-sì.',
-	'wva-protected' => "Ës filmà-sì a l'é protegiù.",
-	'wva-thumbnail-upload-failed' => "Anseriment dël filmà falì, a l'é pa podusse pijé la miniadura aproprià",
-);
-
-/** Pashto (پښتو)
- * @author Ahmed-Najib-Biabani-Ibrahimkhel
- */
-$messages['ps'] = array(
-	'wva-success' => 'د $1 ويډيو مخ په برياليتوب سره ورګډ شو.',
-	'wva-success-replaced' => 'د $1 ويډيو مخ په برياليتوب سره ځايناستی شو.',
-	'wva-failure' => 'لطفاً سم پاراميټرونه ورکړۍ.',
-	'wva-name' => 'نوم يې وليکۍ:',
-	'wva-url' => 'URL يې وليکۍ:',
-	'wva-add' => 'ورګډول',
-	'wva-title' => 'په چټکۍ له يو آر اېل (URL) نه ويډيو ورګډول',
-	'wva-notlogged' => 'د ويډيو د ورننوېستلو لپاره پکار ته چې تاسې غونډال ته ننوځۍ او يا هم ځان ثبت کړی.',
-	'wva-notallowed' => 'تاسې په دې ويکي کې ويډيو نه شی ورننوېستلی.',
-	'wva-protected' => 'دا ويډيو ژغورل شوې',
-);
-
-/** Portuguese (português)
- * @author Crazymadlover
- * @author Giro720
- * @author Hamilton Abreu
- * @author SandroHc
- */
-$messages['pt'] = array(
-	'wva-success' => 'A página do vídeo $1 foi adicionada com sucesso.',
-	'wva-success-replaced' => 'A página do vídeo $1 foi substituída com sucesso.',
-	'wva-failure' => 'Forneça parâmetros válidos, por favor.',
-	'wva-name' => 'Introduza o nome:',
-	'wva-url' => 'Introduza a URL:',
-	'wva-add' => 'Adicionar',
-	'wva-title' => 'Adição Rápida de Vídeo a Partir da URL',
-	'wva-notlogged' => 'Precisa de se autenticar ou registar para inserir vídeos.',
-	'wva-notallowed' => 'Não pode inserir vídeos nesta wiki.',
-	'wva-protected' => 'Este vídeo está protegido.',
-	'wva-thumbnail-upload-failed' => 'A incorporação de vídeo falhou, não foi possível obter uma miniatura',
-);
-
-/** Brazilian Portuguese (português do Brasil)
- * @author Daemorris
- * @author Giro720
- */
-$messages['pt-br'] = array(
-	'wva-success' => 'Página de vídeo $1 foi adicionada com sucesso.',
-	'wva-success-replaced' => 'Página de Vídeo $1 foi substituída com sucesso.',
-	'wva-failure' => 'Por favor, forneça parâmetros válidos.',
-	'wva-name' => 'Especifique o nome:',
-	'wva-url' => 'Especifique a URL:',
-	'wva-add' => 'Adicionar',
-	'wva-title' => 'Adicionar Vídeo Rapidamente Via URL',
-	'wva-notlogged' => 'Você precisa se autenticar ou registrar para inserir vídeos.',
-	'wva-notallowed' => 'Você não pode inserir vídeos nesta wiki.',
-	'wva-protected' => 'Este vídeo está protegido.',
-);
-
-/** Romanian (română)
- * @author Minisarm
- * @author Misterr
- * @author Stelistcristi
- */
-$messages['ro'] = array(
-	'wva-success' => 'Clipul video $1 a fost adăugat cu succes.',
-	'wva-name' => 'Introdu numele:',
-	'wva-url' => 'Introduceți adresa URL:',
-	'wva-add' => 'Adaugă',
-	'wva-notlogged' => 'Trebuie să vă autentificați sau să vă înregistrați pentru a putea adăuga clipuri video.',
-	'wva-notallowed' => 'Nu puteți adăuga clipuri video pe acest wiki.',
-	'wva-protected' => 'Acest clip este protejat.',
-);
-
-/** Russian (русский)
- * @author Eleferen
- * @author Kuzura
- * @author Lockal
- * @author Александр Сигачёв
- */
-$messages['ru'] = array(
-	'wva-success' => 'Страница видео $1 успешно добавлена.',
-	'wva-success-replaced' => 'Страница видео $1 успешно заменена.',
-	'wva-failure' => 'Пожалуйста, укажите верные параметры.',
-	'wva-name' => 'Введите имя:',
-	'wva-url' => 'Введите URL:',
-	'wva-add' => 'Добавить',
-	'wva-title' => 'Быстрое добавление видео по URL',
-	'wva-notlogged' => 'Чтобы добавить видео вам необходимо представиться системе или зарегистрироваться.',
-	'wva-notallowed' => 'Вы не можете вставлять видео в этой вики.',
-	'wva-protected' => 'Это видео защищено.',
-	'wva-thumbnail-upload-failed' => 'Видео добавлено с ошибкой, невозможно получить корректной миниатюры',
-);
-
-/** Sinhala (සිංහල) */
-$messages['si'] = array(
-	'wva-name' => 'නම යොදන්න:',
-	'wva-url' => 'URLඑක යොදන්න:',
-	'wva-add' => 'එක්කරන්න',
-	'wva-protected' => 'මෙම වීඩියෝව ආරක්‍ෂිතය',
-);
-
-/** Serbian (Cyrillic script) (‪српски (ћирилица)‬)
- * @author Charmed94
- * @author Rancher
- * @author Verlor
- */
-$messages['sr-ec'] = array(
-	'wva-success' => 'Видео-страница $1 је додата.',
-	'wva-success-replaced' => 'Видео-страница $1 је замењена.',
-	'wva-failure' => 'Унесите исправне параметре.',
-	'wva-name' => 'Унесите име:',
-	'wva-url' => 'Унесите URL адресу:',
-	'wva-add' => 'Додај',
-	'wva-title' => 'Брзо додавање видео-снимка с адресе',
-	'wva-notlogged' => 'Треба да будете пријављени или да отворите налог да бисте убацивали видео-снимке.',
-	'wva-notallowed' => 'Не можете да убацујете видео-снимке на овај вики.',
-	'wva-protected' => 'Овај видео-снимак је заштићен.',
-);
-
-/** Swedish (svenska)
- * @author Tobulos1
- * @author WikiPhoenix
- */
-$messages['sv'] = array(
-	'wva-success' => 'Video-sidan $1 har lagts till.',
-	'wva-success-replaced' => 'Video-sidan $1 har blivit ersatt.',
-	'wva-failure' => 'Vänligen ange giltiga parametrar.',
-	'wva-name' => 'Ange namn:',
-	'wva-url' => 'Ange URL:',
-	'wva-add' => 'Lägg till',
-	'wva-title' => 'Lägg till video snabbt från URL',
-	'wva-notlogged' => 'Du måste logga in eller registrera dig för att infoga videoklipp.',
-	'wva-notallowed' => 'Du kan inte infoga videoklipp på denna wiki.',
-	'wva-protected' => 'Denna video är skyddad.',
-	'wva-thumbnail-upload-failed' => 'Inbäddningen av videoklippet misslyckades; det gick inte att hämta en miniatyr',
-);
-
-/** Tamil (தமிழ்)
- * @author மதனாஹரன்
- */
-$messages['ta'] = array(
-	'wva-success' => 'காணொளிப் பக்கம் $1 வெற்றிகரமாகச் சேர்க்கப்பட்டது.',
-	'wva-name' => 'பெயரை உள்ளிடவும்:',
-	'wva-url' => 'உரலியை உள்ளிடவும்:',
-	'wva-add' => 'சேர்க்கவும்',
-	'wva-title' => 'உரலியிலிருந்து விரைந்து காணொளியைச் சேர்க்கவும்',
-	'wva-notlogged' => 'காணொளிகளைச் சேர்ப்பதற்கு நீங்கள் உள்நுழைய அல்லது பதிவு செய்ய வேண்டும்.',
-	'wva-notallowed' => 'இவ்விக்கியில் நீங்கள் காணொளிகளைச் சேர்க்க முடியாது.',
-	'wva-protected' => 'இக்காணொளி காக்கப்பட்டுள்ளது.',
-);
-
-/** Telugu (తెలుగు)
- * @author Veeven
- */
-$messages['te'] = array(
-	'wva-add' => 'చేర్చు',
-);
-
-/** Tagalog (Tagalog)
- * @author AnakngAraw
- */
-$messages['tl'] = array(
-	'wva-success' => 'Matagumpay na naidagdag ang pahinang pambidyo na $1.',
-	'wva-success-replaced' => 'Matagumpay na napalitan ang pahinang pambidyo na $1.',
-	'wva-failure' => 'Pakibigay ang tanggap na mga parametro.',
-	'wva-name' => 'Ipasok ang pangalan:',
-	'wva-url' => 'Ipasok ang URL:',
-	'wva-add' => 'Idagdag',
-	'wva-title' => 'Mabilisang Idagdag ang Bidyo Mula sa URL',
-	'wva-notlogged' => 'Kailangan mong lumagda o magpatala upang makapagsingit ng mga bidyo.',
-	'wva-notallowed' => 'Hindi ka makapagsisingit ng mga bidyo sa wiking ito.',
-	'wva-protected' => 'Ang bidyo ay nakaprutekta.',
-	'wva-thumbnail-upload-failed' => 'Nabigo ang pagbabaon ng bidyo, hindi nagawang ang pagkuha ng naaangkop na kagyat',
-);
-
-/** толышә зывон (толышә зывон)
- * @author Гусейн
- */
-$messages['tly'] = array(
-	'wva-add' => 'Зијод кардеј',
-);
-
-/** Ukrainian (українська)
- * @author Ast
- * @author Prima klasy4na
- * @author Тест
- */
-$messages['uk'] = array(
-	'wva-success' => 'Сторінка відео $1 була успішно записана.',
-	'wva-success-replaced' => 'Сторінка відео $1 була успішно замінена.',
-	'wva-failure' => 'Будь ласка, вкажіть вірні параметри.',
-	'wva-name' => 'Введіть назву:',
-	'wva-url' => 'Введіть URL:',
-	'wva-add' => 'Додати',
-	'wva-title' => 'Швидке додавання відео по URL',
-	'wva-notlogged' => 'Для того, щоб вставити відео, Вам потрібно увійти до системи або зареєструватися.',
-	'wva-notallowed' => 'Ви не можете додавати відео у цю Вікі.',
-	'wva-protected' => 'Це відео захищено.',
-);
-
-/** Simplified Chinese (‪中文（简体）‬)
- * @author Anakmalaysia
- * @author Hydra
- */
-$messages['zh-hans'] = array(
-	'wva-name' => '输入名字：',
-	'wva-url' => '输入URL：',
-	'wva-add' => '添加',
-);
-
